% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-BIC.R
\name{BIC}
\alias{BIC}
\title{Bayesian Information Criterion}
\usage{
BIC(hp_0, hp_i, hp_k, db, kern_0, kern_i, m_k, mu_k_param, clust = T)
}
\arguments{
\item{hp_0}{A tibble, data frame or name vector of hyper-parameters for your initial hyperparameters.}

\item{hp_i}{A tibble, data frame or name vector of hyper-parameters at corresponding inputs.}

\item{hp_k}{A tibble, data frame or name vector of hyper-parameters at corresponding variations.}

\item{db}{A tibble containing values we want to compute logL on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{kern_0}{Kernel used to compute the covariance matrix of the mean GP at corresponding inputs (K_0).}

\item{kern_i}{Kernel used to compute the covariance matrix of individuals GP at corresponding inputs (Psi_i).}

\item{m_k}{prior value of the mean parameter of the mean GPs (mu_k). Length = 1 or nrow(db).}

\item{mu_k_param}{parameters of the variational distributions of mean GPs (mu_k).}

\item{clust}{A Boolean. If True \link{...} if False \link{...}.}
}
\value{
Variational BIC
}
\description{
Bayesian Information Criterion
}
