% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-to-matrices.R
\name{check_symmetric}
\alias{check_symmetric}
\title{Round a matrix to make if symmetric}
\usage{
check_symmetric(mat, digits = 10)
}
\arguments{
\item{mat}{A matrix, possibly non-symmetric.}

\item{digits}{A number, the starting number of digits to round from if
\code{mat} is not symmetric}
}
\value{
A matrix, rounded approximation of \code{mat} that is symmetric.
}
\description{
If a matrix is non-symmetric due to numerical errors, round with a decreasing
number of digits until the matrix becomes symmetric.
}
\examples{
TRUE
}
\keyword{internal}
