% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-to-matrices.R
\name{chol_inv_jitter}
\alias{chol_inv_jitter}
\title{Inverse a matrix using an adaptive jitter term}
\usage{
chol_inv_jitter(mat, pen_diag)
}
\arguments{
\item{mat}{A matrix, possibly singular.}

\item{pen_diag}{A number, a jitter term to add on the diagonal.}
}
\value{
A matrix, inverse of \code{mat} plus an adaptive jitter term
added on the diagonal.
}
\description{
Inverse a matrix from its Choleski decomposition. If (nearly-)singular,
increase the order of magnitude of the jitter term added to the diagonal
until the matrix becomes non-singular.
}
\examples{
TRUE
}
