% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{coerce_grid_to_df}
\alias{coerce_grid_to_df}
\title{Coerce a grid object into a tibble}
\usage{
coerce_grid_to_df(grid_obj)
}
\arguments{
\item{grid_obj}{A numeric vector or data.frame representing the input grid.}
}
\value{
A \code{tibble} containing the grid data.
}
\description{
Ensures that the provided grid object is consistently formatted as a tibble.
If the input is a simple numeric vector, it is converted to a tibble with a single
column named 'Input'.
}
\keyword{internal}
