% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{convolution_kernel}
\alias{convolution_kernel}
\title{Compute the Covariance Matrix for a Multi-Output GP via Convolution}
\usage{
convolution_kernel(x, y, hp, vectorized = FALSE, deriv = NULL)
}
\arguments{
\item{x}{The input data. For the vectorized multi-output case, this must be
a tibble/data.frame containing the coordinates and an 'Output_ID' column.}

\item{y}{The second input data (must have the same format as x).}

\item{hp}{For the vectorized multi-output case, this must be a tibble
containing the hyperparameters 'l_t', 'S_t', 'l_u_t' and an 'Output_ID'
column.}

\item{vectorized}{If TRUE, enables the calculation of the full MO covariance
matrix.}

\item{deriv}{A character string specifying the partial derivative to compute.
Names embed the Output_ID label: 'l_t_\if{html}{\out{<o>}}', 'S_t_\if{html}{\out{<o>}}', 'l_u_t' (single}
}
\value{
The covariance matrix or its partial derivative.
}
\description{
Compute the Covariance Matrix for a Multi-Output GP via Convolution
}
