% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{data_allocate_cluster}
\alias{data_allocate_cluster}
\title{Allocate training data into the most probable cluster}
\usage{
data_allocate_cluster(trained_model)
}
\arguments{
\item{trained_model}{A list, containing  the information coming from a
MagmaClust model, previously trained using the
\code{\link{train_magmaclust}} function.}
}
\value{
The original dataset used to train the MagmaClust model, with
additional 'Cluster' and associated 'Proba' columns, indicating the most
probable cluster for each individual/task at the end of the training
procedure.
}
\description{
Allocate training data into the most probable cluster
}
\examples{
TRUE
}
