% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{dmnorm}
\alias{dmnorm}
\title{Compute the Multivariate Gaussian likelihood}
\usage{
dmnorm(x, mu, inv_Sigma, log = FALSE)
}
\arguments{
\item{x}{A vector, containing values the likelihood is evaluated on.}

\item{mu}{A vector or matrix, specifying the mean parameter.}

\item{inv_Sigma}{A matrix, specifying the inverse of covariance parameter.}

\item{log}{A logical value, indicating whether we return the log-likelihood.}
}
\value{

}
\description{
Modification of the function \code{dmnorm()} from the package \code{mvtnorm},
providing an implementation of the Multivariate Gaussian likelihood. This
version uses inverse of the covariance function as argument instead of the
traditional covariance.
}
\examples{
dmnorm(c(1, 2), c(0,0), cbind(c(1,0), c(0,1)), F)
dmnorm(c(1, 2), c(0,0), cbind(c(1,0), c(0,1)), T)
}
