% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{draw}
\alias{draw}
\title{Draw a number}
\usage{
draw(int)
}
\arguments{
\item{int}{An interval of values we want to draw uniformly in.}
}
\value{
A 2-decimals-rounded random number
}
\description{
Draw uniformly a number within a specified interval
}
\examples{
draw(c(1, 2))
}
