% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-em-algorithm.R
\name{e_step_VEM}
\alias{e_step_VEM}
\title{Variation of the E-Step of the EM algorithm}
\usage{
e_step_VEM(db, m_k, kern_0, kern_i, hp_k, hp_i, old_tau_i_k, pen_diag = NULL)
}
\arguments{
\item{db}{A tibble or data frame. Columns required: ID, Input, Output.
Additional columns for covariates can be specified.}

\item{m_k}{prior means of the mu_k processes.}

\item{kern_0}{A kernel function, associated with the mean GP.}

\item{kern_i}{A kernel function, associated with the individual GPs.}

\item{hp_k}{A named vector, tibble or data frame of hyper-parameters
associated with \code{kern_0}.}

\item{hp_i}{A named vector, tibble or data frame of hyper-parameters
associated with \code{kern_i}.}

\item{old_tau_i_k}{values au tau_i_k from previous iterations. List(list(tau)_i)_k}

\item{pen_diag}{A number. A jitter term, added on the diagonal to prevent
numerical issues when inverting nearly singular matrices.}
}
\value{
A named list, containing the elements \code{mean}, a tibble
containing the Input and associated Output of the hyper-posterior mean
parameter, and \code{cov}, the hyper-posterior covariance matrix.

A named list, containing the elements \code{mean}, a tibble
containing the Input and associated Output of the hyper-posterior mean
parameter, and \code{cov}, the hyper-posterior covariance matrix.
}
\description{
Expectation step of the variation of the EM algorithm to compute the parameters of the
hyper-posterior Gaussian distribution of the mean process in Magma.
}
\examples{
\donttest{
k = seq_len(3)
m_k <- c("K1" = 0, "K2" = 0, "K3" = 0)

db <- simu_db(N = 10, common_input = TRUE)
hp_k <- MagmaClustR:::hp("SE", list_ID = names(m_k))
hp_i <- MagmaClustR:::hp("SE", list_ID = unique(db$ID))

old_tau_i_k = MagmaClustR:::ini_tau_i_k(db = db, k = length(k), nstart = 50)
pi_1 <- old_tau_i_k \%>\% dplyr::select(-.data$ID)
hp_k[['pi']] = sapply( pi_1, function(x) x \%>\% unlist() \%>\% mean() )

MagmaClustR:::e_step_VEM(db, m_k, "SE", "SE", hp_k, hp_i, old_tau_i_k ,0.001)

}

}
