% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbos.R
\name{elbo_clust_multi_GP}
\alias{elbo_clust_multi_GP}
\title{Evidence Lower Bound for a mixture of GPs}
\usage{
elbo_clust_multi_GP(hp, db, hyperpost, kern, pen_diag)
}
\arguments{
\item{hp}{A tibble, data frame or named vector containing hyper-parameters.}

\item{db}{A tibble containing the values we want to compute the elbo on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{hyperpost}{List of parameters for the K mean GPs.}

\item{kern}{A kernel function used to compute the covariance matrix at
corresponding timestamps.}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
The value of the penalised Gaussian elbo for a mixture of GPs
}
\description{
Evidence Lower Bound for a mixture of GPs
}
\examples{
TRUE
}
\keyword{internal}
