% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbos.R
\name{elbo_clust_multi_GP_shared_hp_i}
\alias{elbo_clust_multi_GP_shared_hp_i}
\title{Penalised elbo for multiple individual GPs with common HPs}
\usage{
elbo_clust_multi_GP_shared_hp_i(hp, db, hyperpost, kern, pen_diag)
}
\arguments{
\item{hp}{A tibble, data frame or named vector containing hyper-parameters.}

\item{db}{A tibble containing values we want to compute elbo on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{hyperpost}{List of parameters for the K mean Gaussian processes.}

\item{kern}{A kernel function used to compute the covariance matrix at
corresponding timestamps.}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
The value of the penalised Gaussian elbo for
the sum of the M individual GPs with common HPs.
}
\description{
Penalised elbo for multiple individual GPs with common HPs
}
\examples{
TRUE
}
\keyword{internal}
