% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-MagmaClust.R
\name{elbo_GP_mod_common_hp_k}
\alias{elbo_GP_mod_common_hp_k}
\title{Modified Gaussian log-likelihood for the sum of the k mean GPs with same HPs}
\usage{
elbo_GP_mod_common_hp_k(hp, db, mean, kern, post_cov, pen_diag = NULL)
}
\arguments{
\item{hp}{A tibble, data frame or name vector containing hyper-parameters.}

\item{db}{A tibble containing values we want to compute elbo on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{mean}{list of the k means of the GP at union of observed timestamps}

\item{kern}{A kernel function used to compute the covariance matrix at corresponding timestamps.}

\item{post_cov}{A List of the k posterior covariance of the mean GP (mu_k). Used to compute correction term (cor_term).}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
value of the modified Gaussian log-likelihood for the sum of the k mean GPs with same HPs
}
\description{
Modified Gaussian log-likelihood for the sum of the k mean GPs with same HPs
}
