% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{expand_grid_inputs}
\alias{expand_grid_inputs}
\title{Expand a grid of inputs}
\usage{
expand_grid_inputs(Input, ...)
}
\arguments{
\item{Input}{A vector of inputs.}

\item{...}{As many vector of covariates as desired. We advise to give
explicit names when using the function.}
}
\value{
A tibble containing all the combination of values of the
parameters.
}
\description{
Expand a grid of inputs
}
\examples{
TRUE
}
