% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{format_longer}
\alias{format_longer}
\title{Pivot MagmaClustR data to long format}
\usage{
format_longer(db)
}
\arguments{
\item{db}{A data frame or tibble containing the simulated data.}
}
\value{
A formatted \code{tibble} in long format.
}
\description{
Converts a simulated dataset from a wide format (e.g. containing an 'ID',
'Input_1', 'Input_2', 'Output') to a standardized long format with
'Task_ID', 'Input_ID', 'Input', 'Output_ID', and 'Output'.
}
