% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{format_wider}
\alias{format_wider}
\title{Pivot MagmaClustR data to wide format}
\usage{
format_wider(db)
}
\arguments{
\item{db}{A data frame or tibble containing the simulated data in long format.}
}
\value{
A formatted \code{tibble} in wide format.
}
\description{
Converts a simulated dataset from the standard long format back to a
wider format, extracting dimensions into 'Input_1', 'Input_2', etc.,
and renaming 'Task_ID' to 'ID'.
}
