% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma.R
\name{full_algo}
\alias{full_algo}
\title{Full algorithm MAGMA}
\usage{
full_algo(
  db,
  new_db,
  timestamps,
  kern_i,
  common_hp = T,
  plot = T,
  prior_mean,
  kern_0 = NULL,
  list_hp = NULL,
  mu = NULL,
  ini_hp_0 = NULL,
  ini_hp_i = NULL,
  hp_new_i = NULL
)
}
\arguments{
\item{db}{Database containing all training data from all individuals. Column: ID - input - Output.}

\item{new_db}{Database containing data for a new individual we want a prediction on.}

\item{timestamps}{Timestamps we want to predict at.}

\item{kern_i}{Kernel associated to individual GPs.}

\item{common_hp}{your common Hyperparameters}

\item{plot}{Boolean indicating whether we want to display a graph at the end of computations.}

\item{prior_mean}{Prior arbitrary value for the mean process. Optional, not needed if 'mu' is given.}

\item{kern_0}{Kernel associated to the mean GPs. Optional, not needed if 'mu' is given.}

\item{list_hp}{Hyper-parameters for all individuals in training set. Optional, computed if NULL.}

\item{mu}{Database containing parameters of the mean GP at all prediction timestamps. Optional, computed if NULL.}

\item{ini_hp_0}{Initial values of the HP to start the training for corresponding input. Optional, not needed if 'list_hp' is given.}

\item{ini_hp_i}{Initial values of the HP to start the training for individual. Optional, not needed if 'list_hp' is given.}

\item{hp_new_i}{Hyper-pameters for the new individual to predict. Optional, computed if NULL.}
}
\value{
predicted GP parameters | posterior mean process | all trained hyperparameters
}
\description{
Full algorithm MAGMA
}
\examples{
TRUE
}
