% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magmaClust.R
\name{full_algo_clust}
\alias{full_algo_clust}
\title{Full algorithm MAGMACLUST}
\usage{
full_algo_clust(
  db,
  new_db,
  timestamps,
  kern_i,
  ini_tau_i_k = NULL,
  common_hp_k = T,
  common_hp_i = T,
  prior_mean = NULL,
  kern_0 = NULL,
  list_hp = NULL,
  mu_k = NULL,
  ini_hp_k = NULL,
  ini_hp_i = NULL,
  hp_new_i = NULL
)
}
\arguments{
\item{db}{Database containing all training data from all individuals. Column: ID - Input - Output.}

\item{new_db}{Database containing data for a new individual we want a prediction on.}

\item{timestamps}{Timestamps we want to predict at.}

\item{kern_i}{Kernel associated to individual GPs.}

\item{ini_tau_i_k}{Initial probability to belong to each cluster for individuals. Optional, if list_hp is given}

\item{common_hp_k}{boolean indicating whether hp are common among mean GPs (for each mu_k)}

\item{common_hp_i}{boolean indicating whether hp are common among individual GPs (for each y_i)}

\item{prior_mean}{Prior arbitrary value for the mean process. Optional, not needed if 'mu' is given.}

\item{kern_0}{Kernel associated to the mean GPs. Optional, not needed if 'mu' is given.}

\item{list_hp}{Hyper-parameters for all individuals in training set. Optional, computed if NULL.}

\item{mu_k}{list containing parameters of the mean GPs at all prediction timestamps. Optional, computed if NULL.}

\item{ini_hp_k}{named vector, tibble or data frame of hyper-parameters
associated with \code{kern_0}, the mean process' kernel. The
columns/elements should be named according to the hyper-parameters
that are used in \code{kern_0}.}

\item{ini_hp_i}{A tibble or data frame of hyper-parameters
associated with \code{kern_i}, the individual processes' kernel.
Required column : \code{ID}. The \code{ID} column contains the unique
names/codes used to identify each individual/task. The other columns
should be named according to the hyper-parameters that are used in
\code{kern_i}.#' @param hp_new_i Hyper-pameters for the new individual to predict. Optional, computed if NULL.}

\item{hp_new_i}{Hyper-pameters for the new individual to predict. Optional, computed if NULL.}
}
\value{
predicted GP parameters | posterior mean process | all trained hyperparameters


}
\description{
Full algorithm MAGMACLUST
}
