% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{generate_mean_process}
\alias{generate_mean_process}
\title{Generate a Mean Process Realization (Convolution Kernel Version)
Creates a working grid for each output and generates a single smooth
underlying mean process, \code{mu_0}, by drawing a realization from a multi-output
Gaussian Process using a convolution kernel.}
\usage{
generate_mean_process(
  points_per_output,
  grid_input,
  hp_config_mean_process,
  prior_means,
  noise_0 = NULL,
  shared_grid_outputs,
  shared_hp_outputs,
  shared_grid_clusters = FALSE,
  precomputed_grid_list = NULL,
  v = NULL,
  l = NULL,
  sigma = 0
)
}
\arguments{
\item{points_per_output}{A numeric vector specifying the number of points for
each output's working grid.}

\item{grid_input}{A list of numeric vectors, where each defines the
input domain for the corresponding output.}

\item{hp_config_mean_process}{A tibble configuring the hyperparameters for
the mean process's convolution kernel.}

\item{prior_means}{A vector containing the values of the prior mean of each
output.}

\item{noise_0}{An optional numeric vector specifying the standard deviation
of the noise for each output. If provided, its length must
match the number of outputs. Default is \code{NULL} (no noise).}

\item{shared_grid_outputs}{A logical value. If \code{TRUE}, all outputs are
defined on the exact same input grid.}

\item{shared_hp_outputs}{A logical value. If \code{TRUE}, all outputs share the
same hyperparameters.}

\item{shared_grid_clusters}{A logical value. If \code{TRUE}, clusters share the same grid.}

\item{precomputed_grid_list}{An optional list of precomputed grids, one per output.}

\item{v}{A number. The variance hyper-parameter of the SE kernel.}

\item{l}{A number. The lengthscale hyper-parameter of the SE kernel.}

\item{sigma}{A number. The noise hyper-parameter.}
}
\value{
A list containing the realization, grid, and hyperparameters.
}
\description{
Generate a Mean Process Realization (Convolution Kernel Version)
Creates a working grid for each output and generates a single smooth
underlying mean process, \code{mu_0}, by drawing a realization from a multi-output
Gaussian Process using a convolution kernel.
}
\keyword{internal}
