% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-likelihoods-MagmaClust.R
\name{gr_clust_multi_GP}
\alias{gr_clust_multi_GP}
\title{Gradient multi-Gaussian Process}
\usage{
gr_clust_multi_GP(hp, db, mu_k_param, kern)
}
\arguments{
\item{hp}{A tibble, data frame or name vector of hyper-parameters.}

\item{db}{A tibble containing values we want to compute logL on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{mu_k_param}{List of parameters for the K mean Gaussian processes.}

\item{kern}{A kernel function used to compute the covariance matrix at corresponding timestamps.}
}
\value{
The gradient of multi gaussian processes for clustering
}
\description{
Gradient multi-Gaussian Process
}
