% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-likelihoods-MagmaClust.R
\name{gr_clust_multi_GP_common_hp_i}
\alias{gr_clust_multi_GP_common_hp_i}
\title{Modified common Gaussian Process for each variations.}
\usage{
gr_clust_multi_GP_common_hp_i(hp, db, mu_k_param, kern, pen_diag = NULL)
}
\arguments{
\item{hp}{A tibble, data frame or name vector of hyper-parameters.}

\item{db}{A tibble containing values we want to compute elbo on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{mu_k_param}{List of parameters for the K mean Gaussian processes.}

\item{kern}{A kernel function used to compute the covariance matrix at
corresponding timestamps.}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
The value of the modified Gaussian log-likelihood for
one GP as it appears in the model.
}
\description{
Modified common Gaussian Process for each variations.
}
\examples{
TRUE
}
