% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-likelihoods.R
\name{gr_GP}
\alias{gr_GP}
\title{Gradient Gaussian Process}
\usage{
gr_GP(hp, db, mean, kern, new_cov)
}
\arguments{
\item{hp}{Set of hyper-parameter.}

\item{db}{Full database with all individuals. Columns required : ID, Input, Output}

\item{mean}{Mean of your Gaussian Process.}

\item{kern}{Kernel used to compute the covariance matrix.}

\item{new_cov}{Posterior covariance matrix of the mean GP (mu_0).}
}
\value{
Gradient of the Gaussian Process
}
\description{
Gradient Gaussian Process
}
