% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-likelihoods.R
\name{gr_GP_mod_common_hp}
\alias{gr_GP_mod_common_hp}
\title{Gradient common Gaussian Process}
\usage{
gr_GP_mod_common_hp(hp, db, mean, kern, new_cov, pen_diag)
}
\arguments{
\item{hp}{set of hyper-parameter}

\item{db}{full database with all individuals. Columns required : ID, input, Output}

\item{mean}{mean of the GP at union of observed input}

\item{kern}{kernel used to compute the covariance matrix at corresponding inputs}

\item{new_cov}{posterior covariance matrix of the mean GP (mu_0).}

\item{pen_diag}{A jitter}
}
\value{
Gradient of the common Gaussian Process
}
\description{
Gradient common Gaussian Process
}
\examples{
TRUE
}
