% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-likelihoods-MagmaClust.R
\name{gr_GP_mod_common_hp_k}
\alias{gr_GP_mod_common_hp_k}
\title{Modified common Gaussian Process for each variations.}
\usage{
gr_GP_mod_common_hp_k(hp, db, mean, kern, new_cov, pen_diag = NULL)
}
\arguments{
\item{hp}{A tibble, data frame or name vector of hyper-parameters.}

\item{db}{A tibble containing values we want to compute logL on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{mean}{mean of the GP at union of observed input}

\item{kern}{A kernel function used to compute the covariance matrix at corresponding timestamps.}

\item{new_cov}{posterior covariance matrix of the mean GP (mu_0).}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{

}
\description{
Modified common Gaussian Process for each variations.
}
