% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{grid_n_points}
\alias{grid_n_points}
\title{Get the number of points in a grid object}
\usage{
grid_n_points(grid_obj)
}
\arguments{
\item{grid_obj}{A numeric vector, list, or data.frame representing the input grid.}
}
\value{
An integer corresponding to the length of the vector or the number of rows
of the data.frame.
}
\description{
Evaluates the size of a grid, whether it is formatted as a simple vector
or a multidimensional data.frame.
}
\keyword{internal}
