% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{ini_mixture}
\alias{ini_mixture}
\title{Mixture initialisation with kmeans}
\usage{
ini_mixture(data, k, name_clust = NULL, nstart = 50)
}
\arguments{
\item{data}{A tibble or data frame. Required columns: \code{ID}, \code{Input}
, \code{Output}.}

\item{k}{A number, indicating the number of clusters.}

\item{name_clust}{A vector of characters. Each element should correspond to
the name of one cluster.}

\item{nstart}{A number of restart used in the underlying kmeans algorithm}
}
\value{
A tibble indicating for each \code{ID} in which cluster it belongs
after a kmeans initialisation.
}
\description{
Provide an initial kmeans allocation of the individuals/tasks in a dataset
into a definite number of clusters, and return the associated mixture
probabilities.
}
\examples{
TRUE
}
