% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{ini_tau_i_k}
\alias{ini_tau_i_k}
\title{ini_tau_i_k}
\usage{
ini_tau_i_k(db, k, nstart = 50)
}
\arguments{
\item{db}{database}

\item{k}{cluster}

\item{nstart}{nstart}
}
\value{
ini_tau_i_k
}
\description{
ini_tau_i_k
}
