% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{is_flat_mo_hp}
\alias{is_flat_mo_hp}
\title{Test whether a hyper-parameter object is a flat multi-output vector}
\usage{
is_flat_mo_hp(hp)
}
\arguments{
\item{hp}{An object to test.}
}
\value{
A logical value.
}
\description{
Test whether a hyper-parameter object is a flat multi-output vector
}
\keyword{internal}
