% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kern_to_cov.R
\name{kern_to_cov}
\alias{kern_to_cov}
\title{Kernel to Covariance}
\usage{
kern_to_cov(input, kern, hp)
}
\arguments{
\item{input}{tibble of all inputs for each individuals.}

\item{kern}{indicates which kernel function to use to compute the covariance matrix}

\item{hp}{the hyperparameters used in your kernel}
}
\value{
list of covariance matrices (1 by individual) of the input vectors according to kernel() function
}
\description{
Kernel to Covariance
}
\examples{
x = rbind(c(1,0,1),c(2,1,2),c(1,2,3))
kern_to_cov(x,kernel_sqrd_exp, tibble::tibble(sigma = 1, lengthscale = 0.5))

}
