% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-to-matrices.R
\name{list_kern_to_cov}
\alias{list_kern_to_cov}
\title{Compute a covariance matrix for multiple individuals}
\usage{
list_kern_to_cov(data, kern, hp, deriv = NULL)
}
\arguments{
\item{data}{A tibble or data frame of input data. Required column: 'ID'.
Suggested column: 'Input' (for indicating the reference input).}

\item{kern}{A kernel function.}

\item{hp}{A tibble or data frame, containing the hyper-parameters associated
with each individual.}

\item{deriv}{A character, indicating according to which hyper-parameter the
derivative should be computed. If NULL (defaut), the function simply returns
the list of covariance matrices.}
}
\value{
A named list containing all of the inverse covariance matrices.
}
\description{
Compute the covariance matrices associated with all individuals in the
database, taking into account their specific inputs and hyper-parameters.
}
\examples{
db <- simu_db(M = 3)
hp <- tibble::tibble(ID = unique(db$ID), hp())
MagmaClustR:::list_kern_to_cov(db, "SE", hp)
}
