% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods-MagmaClust.R
\name{logL_clust_multi_GP}
\alias{logL_clust_multi_GP}
\title{Log-Likelihood function for clusturing mutli-Gaussian Process}
\usage{
logL_clust_multi_GP(hp, db, mu_k_param, kern, pen_diag)
}
\arguments{
\item{hp}{A tibble, data frame or name vector containing hyper-parameters.}

\item{db}{A tibble containing the values we want to compute the logL on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{mu_k_param}{List of parameters for the K mean Gaussian processes.}

\item{kern}{A kernel function used to compute the covariance matrix at corresponding timestamps.}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
The value of the modified Gaussian log-likelihood for one GP as it appears in the model.
}
\description{
Log-Likelihood function for clusturing mutli-Gaussian Process
}
