% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logL_GP}
\alias{logL_GP}
\title{Log Likelihood function of the Gaussian Process}
\usage{
logL_GP(hp, db, mean, kern, new_cov)
}
\arguments{
\item{hp}{The hyperparameters of your kernel}

\item{db}{tibble containing values we want to compute. Required columns : input, Output}

\item{mean}{mean of the GP at corresponding inputs}

\item{kern}{kernel used to compute the covariance matrix at corresponding inputs}

\item{new_cov}{posterior covariance matrix of the mean GP (mu_0). Used to compute correction term (cor_term)}
}
\value{
value of the Gaussian log-likelihood for one GP as it appears in the model
}
\description{
Log Likelihood function of the Gaussian Process
}
\examples{
logL_GP(tibble::tibble(sigma = 1, lengthscale = 0.5),
tibble::tibble(input= 1:5,Output= 2:6),rep(3,5),kernel_sqrd_exp,0)

}
