% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logL_GP_mod_common_hp}
\alias{logL_GP_mod_common_hp}
\title{Modified Gaussian log-likelihood for the sum of all indiv}
\usage{
logL_GP_mod_common_hp(hp, db, mean, kern, new_cov)
}
\arguments{
\item{hp}{vector of common hyperparameters for all individuals.}

\item{db}{tibble of data. Required columns : ID, input, Output}

\item{mean}{mean of the GP at union of observed inputs}

\item{kern}{kernel used to compute the covariance matrix at corresponding inputs}

\item{new_cov}{posterior covariance matrix of the mean GP (mu_0). Used to compute correction term (cor_term)}
}
\value{
value of the modified Gaussian log-likelihood for the sum of all indiv with same HPs
}
\description{
Modified Gaussian log-likelihood for the sum of all indiv
}
\examples{
logL_GP(tibble::tibble(sigma = 1, lengthscale = 0.5),
tibble::tibble(ID = 1:5, input= 1:5,Output= 2:6),rep(3,5),kernel_sqrd_exp,0)
}
