% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logL_monitoring}
\alias{logL_monitoring}
\title{Log-Likelihood for monitoring the EM algorithm}
\usage{
logL_monitoring(hp_0, hp_i, db, kern_i, kern_0, mean_mu, cov_mu, m_0, pen_diag)
}
\arguments{
\item{hp_0}{hyperparameters for the kernel 0}

\item{hp_i}{hyperparameters for each kernel i}

\item{db}{tibble containing values we want to compute logL on. Required columns : ID, input, Output}

\item{kern_i}{kernel used to compute the covariance matrix of individuals GP at corresponding inputs (Psi_i)}

\item{kern_0}{kernel used to compute the covariance matrix of the mean GP at corresponding inputs (K_0)}

\item{mean_mu}{posterior mean of the mean GP (mu_0). Needed to compute the log-likelihood}

\item{cov_mu}{posterior covariance matrix of the mean GP (mu_0). Needed to compute correction term}

\item{m_0}{prior value of the mean parameter of the mean GP (mu_0). Length = 1 or nrow(db)}

\item{pen_diag}{value of the penalization of the diagonal}
}
\value{
value of expectation of joint log-likelihood of the model. The function to be maximised in step M
The full likelihood is composed of M+1 independent parts, depending on only theta_0, or theta_i respectively
for each i. The following code computes and sums these M+1 (modified) gaussian likelihoods.
}
\description{
Log-Likelihood for monitoring the EM algorithm
}
\examples{
kern_i <- kern_0 <- kernel_sqrd_exp
hp_0 <- tibble::tibble(sigma = 1, lengthscale = 0.5)
hp_i <- list(hp_0, hp_0, hp_0, hp_0, hp_0)
db <- mean_mu <- tibble::tibble(ID = 1:5, input = 1:5, Output = 2:6)
}
