% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{make_task_id}
\alias{make_task_id}
\title{Build a task or cluster identifier}
\usage{
make_task_id(i, k, n_clusters)
}
\arguments{
\item{i}{An integer, the task index within its cluster.}

\item{k}{An integer, the cluster index.}

\item{n_clusters}{An integer, the total number of clusters being simulated.}
}
\value{
A character string identifying the task.
}
\description{
Build a task or cluster identifier
}
\keyword{internal}
