% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-BIC.R
\name{model_selection}
\alias{model_selection}
\title{Selection of the model}
\usage{
model_selection(
  db,
  k_grid = 1:5,
  ini_hp_k,
  ini_hp_i,
  kern_0,
  kern_i,
  ini_tau_i_k = NULL,
  common_hp_k = T,
  common_hp_i = T,
  plot = T,
  pen_diag
)
}
\arguments{
\item{db}{A tibble containing values we want to compute logL on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{k_grid}{A vector, indicating the grid of additional reference
inputs on which the mean process' hyper-posterior should be evaluated.}

\item{ini_hp_k}{named vector, tibble or data frame of hyper-parameters
associated with \code{kern_0}, the mean process' kernel. The
columns/elements should be named according to the hyper-parameters
that are used in \code{kern_0}.}

\item{ini_hp_i}{A tibble or data frame of hyper-parameters
associated with \code{kern_i}, the individual processes' kernel.
Required column : \code{ID}. The \code{ID} column contains the unique
names/codes used to identify each individual/task. The other columns
should be named according to the hyper-parameters that are used in
\code{kern_i}.}

\item{kern_0}{A kernel function, associated with the mean GP.}

\item{kern_i}{A kernel function, associated with the individual GPs.}

\item{ini_tau_i_k}{initial values of probabiliy to belong to each cluster for each individuals.}

\item{common_hp_k}{boolean indicating whether hp are common among mean GPs (for each mu_k)}

\item{common_hp_i}{boolean indicating whether hp are common among individual GPs (for each y_i)}

\item{plot}{boolean indicating whether you want to plot or not}

\item{pen_diag}{A number. A jitter term, added on the diagonal to prevent
numerical issues when inverting nearly singular matrices.}
}
\value{
TRUE
}
\description{
Selection of the model
}
\examples{
TRUE
}
