% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-plot-functions.R
\name{plot_animate_clust}
\alias{plot_animate_clust}
\title{plot_animate_clust}
\usage{
plot_animate_clust(
  pred_gp,
  ygrid,
  data = NULL,
  data_train = NULL,
  mean_k = NULL,
  file = "gganim.gif"
)
}
\arguments{
\item{pred_gp}{pred_gp}

\item{ygrid}{ygrid}

\item{data}{data}

\item{data_train}{data_train}

\item{mean_k}{mean_k}

\item{file}{file}
}
\value{
TRUE
}
\description{
plot_animate_clust
}
