% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_db}
\alias{plot_db}
\title{Plot raw data}
\usage{
plot_db(data, cluster = FALSE, legend = FALSE)
}
\arguments{
\item{data}{A data frame or tibble with format : ID, Input, Output for
single output configurations; Task_ID, Input_ID, Input, Output_ID, Output
for multi-output configurations.}

\item{cluster}{A boolean indicating whether data should be coloured by
cluster. Requires a column named 'Cluster'.}

\item{legend}{A boolean indicating whether the legend should be displayed.}
}
\value{
Graph of smoothed curves of raw data.
}
\description{
Display raw data under the Magma format as dot and lines.
}
\examples{
TRUE
}
