% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_gif}
\alias{plot_gif}
\title{Plot GIF of a Magma prediction}
\usage{
plot_gif(
  pred_gp,
  data = NULL,
  data_train = NULL,
  mean = NULL,
  mean_CI = F,
  file = "gganim.gif"
)
}
\arguments{
\item{pred_gp}{tibble coming out of the pred_gp() function, columns required : 'input', 'Mean', 'Var'}

\item{data}{tibble of observational data, columns required : 'input', 'Output' (Optional)}

\item{data_train}{tibble of training data, columns required : 'ID', input', 'Output' (Optional)}

\item{mean}{tibble of estimated value for the mean process, columns required : 'input', 'Mean' (Optional)}

\item{mean_CI}{boolean indicating whether the mean curve should be displayed along with CI95 band or not (Optional)}

\item{file}{The animation file}
}
\value{
plot the animated curves of the GP (GIF) with the 0.95 confidence interval (optional display raw data)
}
\description{
Create an animation displaying as GIF how the
}
\examples{
TRUE
}
