% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_heat}
\alias{plot_heat}
\title{Title}
\usage{
plot_heat(
  pred_gp,
  data = NULL,
  data_train = NULL,
  mean = NULL,
  ygrid = NULL,
  interactive = F,
  CI = T
)
}
\arguments{
\item{pred_gp}{tibble coming out of the pred_gp() function, columns required : 'input', 'Mean', 'Var'}

\item{data}{tibble of observational data for the new individual, columns required : 'input', 'Output' (optional)}

\item{data_train}{tibble of training data for the model, , columns required : 'input', 'Output' (optional)}

\item{mean}{(optional)}

\item{ygrid}{grid on y-axis to plot the heatmap on (optional, default = range of data +- 2 sd)}

\item{interactive}{boolean indicating whether the output plot should be interactive (plotly)}

\item{CI}{boolean. If True, display the heatmap of Credible Intervals. If False, display the heatmap of likelihood}
}
\value{
A heatmap illustrating the GP predition results (Optionnal diplay of raw data and mean process)
}
\description{
Title
}
\examples{
TRUE
}
