% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{posterior_mu}
\alias{posterior_mu}
\title{Posterior mu}
\usage{
posterior_mu(db, new_db, timestamps, m_0, kern_0, kern_i, hp_0, hp_i)
}
\arguments{
\item{db}{matrix of data columns required ('input', 'Output')}

\item{new_db}{Database containing data for a new individual we want a prediction on.}

\item{timestamps}{timestamps on which we want a prediction}

\item{m_0}{prior mean value of the mean GP (scalar value or vector of same length as 'timestamps')}

\item{kern_0}{kernel associated to the covariance function of the mean GP}

\item{kern_i}{Kernel associated to individual GPs.}

\item{hp_0}{Set of hyperparameters of the mean GP}

\item{hp_i}{Set of hyperparameters for each indicidual GPs}
}
\value{
pamameters of the mean GP at timestamps
}
\description{
Posterior mu
}
