% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_magma_animate}
\alias{pred_magma_animate}
\title{Prediction Gaussian Process Animate}
\usage{
pred_magma_animate(db, timestamps = NULL, mean_mu = 0, cov_mu = NULL, kern, hp)
}
\arguments{
\item{db}{Your database, the same as for a classic GP prediction}

\item{timestamps}{dl}

\item{mean_mu}{mean value of mean GP at timestamps (obs + pred)}

\item{cov_mu}{covariance of mean GP at timestamps (obs + pred)}

\item{kern}{kernel of your choice use with your hyperparameters}

\item{hp}{list of hyperparameters of your model}
}
\value{
tibble of classic GP predictions but with an inscreasing number of data points considered as 'observed'
}
\description{
Function used to generate data compatible with a GIF ploting of the results
}
\examples{
TRUE
}
