% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{pred_max_cluster}
\alias{pred_max_cluster}
\title{Prediction of the maximum of cluster}
\usage{
pred_max_cluster(tau_i_k)
}
\arguments{
\item{tau_i_k}{tau_i_k}
}
\value{
Prediction of the maximum of cluster
}
\description{
Prediction of the maximum of cluster
}
