% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{regularize_data}
\alias{regularize_data}
\title{Regularise a grid of inputs in a dataset}
\usage{
regularize_data(
  data,
  size_grid = 30,
  scale = FALSE,
  grid_inputs = NULL,
  summarise_fct = base::mean
)
}
\arguments{
\item{data}{A tibble or data frame. Required columns: \code{ID},
\code{Output}. The \code{ID} column contains the unique names/codes used
to identify each individual/task (or batch of data). The \code{Output}
column specifies the observed values (the response variable). The data
frame can also provide as many inputs as desired, with no constraints
on the column names.}

\item{size_grid}{A number, which indicates how many points each column of the
grid must contain. This number is used when no inputs grid is specified by
the user. See 'grid_inputs' below for details. Default is 30.}

\item{scale}{A boolean. If TRUE, the dataset is scaled to zero mean and
unit variance in all column. Default is FALSE.}

\item{grid_inputs}{A data frame, corresponding to a pre-defined grid of
inputs according to which we want to regularise a dataset. If
NULL (default), a dedicated grid of inputs is defined: for each
input column, a regular sequence is created from the min of the input
values to the max, with a step equal to the 'size_grid' parameter.}

\item{summarise_fct}{A character string or a function. If several similar
inputs are associated with different outputs, the user can choose the
summarising function for the output among the following: min, max, mean,
median. A custom function can be defined if necessary. Default is "mean".}
}
\value{
A data frame, where input columns have been regularised accordingly.
}
\description{
Modify the original grid of inputs to make it more 'regular' (in the sense
that the interval between each observation is constant, or corresponds to a
specific pattern defined by the user). In particular, this function can also
be used to summarise several data points into one, at a specific location. In
this case, the output values are averaged according to the 'summarise_fct'
argument.
}
\examples{
TRUE
}
