% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rq_kernel}
\alias{rq_kernel}
\title{Rational Quadratic Kernel}
\usage{
rq_kernel(x, y, hp)
}
\arguments{
\item{x}{a vector/matrix}

\item{y}{a vector/matrix}

\item{hp}{a tibble of the hyperparameters}
}
\value{
The value of dot production <f(t1),f(t2)> computed from the kernel
}
\description{
Rational Quadratic Kernel
}
\examples{
kernel_quad(c(1,0),c(0,1),tibble::tibble(sigma = 1, lengthscale = 0.5, alpha = 3))
}
