% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_gp}
\alias{sample_gp}
\alias{sample_magma}
\title{Draw samples from a posterior GP/Magma distribution}
\usage{
sample_gp(pred_gp, nb_samples = 50)

sample_magma(pred_gp, nb_samples = 50)
}
\arguments{
\item{pred_gp}{A list, typically coming from
\code{\link{pred_magma}} or \code{\link{pred_gp}} functions, with argument
'get_full_cov = TRUE'. Required elements: \code{pred}, \code{cov}.}

\item{nb_samples}{A number, indicating the number of samples to be drawn from
the predictive posterior distribution. For two-dimensional graphs, only
one sample can be displayed.}
}
\value{
A tibble or data frame, containing the samples generated from
a GP prediction. Format: \code{Input}, \code{Sample}, \code{Output}.
}
\description{
Draw samples from a posterior GP/Magma distribution
}
\examples{
TRUE
}
