% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{sample_grid_points}
\alias{sample_grid_points}
\title{Sample and sort points from a grid object}
\usage{
sample_grid_points(grid_obj, size)
}
\arguments{
\item{grid_obj}{A numeric vector or data.frame representing the input grid.}

\item{size}{An integer specifying the number of points to randomly sample.}
}
\value{
A sorted grid object of the same class as \code{grid_obj}, containing
the sampled points.
}
\description{
Randomly selects a specified number of points from a grid without replacement,
and returns them in a sorted order. It automatically handles both 1D vectors
and multi-dimensional data.frames.
}
\keyword{internal}
