% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_magmaclust}
\alias{sample_magmaclust}
\title{Draw samples from a MagmaClust posterior distribution}
\usage{
sample_magmaclust(pred_clust, nb_samples = 50)
}
\arguments{
\item{pred_clust}{A list, typically coming from
\code{\link{pred_magmaclust}}, with argument get_full_cov = TRUE'.
Required elements: \code{pred}, \code{cov}, \code{mixture}.}

\item{nb_samples}{A number, indicating the number of samples to be drawn from
the predictive posterior distribution. For two-dimensional graphs, only
one sample can be displayed.}
}
\value{
A tibble or data frame, containing the samples generated from
a GP prediction. Format: \code{Cluster}, \code{Proba}, \code{Input},
\code{Sample}, \code{Output}.
}
\description{
Draw samples from a MagmaClust posterior distribution
}
\examples{
TRUE
}
