% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_kernel.R
\name{se_kernel}
\alias{se_kernel}
\title{Squared Exponential Kernel}
\usage{
se_kernel(x, y, hp)
}
\arguments{
\item{x}{a vector/matrix}

\item{y}{a vector/matrix}

\item{hp}{a tibble of the hyperameters}
}
\value{
The value of dot production <f(t1),f(t2)> computed from the kernel
}
\description{
Squared Exponential Kernel
}
\examples{
se_kernel(c(1,0),c(0,1), tibble::tibble(sigma = 1, lengthscale = 0.5))
}
