% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simu_data}
\alias{simu_data}
\alias{simu_db}
\title{Simulate a dataset tailored for MagmaClustR}
\usage{
simu_data(
  n_tasks = 10,
  n_points = 10,
  n_clusters = 1,
  n_outputs = 1,
  input2 = FALSE,
  prior_means = NULL,
  grid_input = seq(-1, 1, 0.01),
  grid_input2 = seq(-1, 1, 0.5),
  shared_input = FALSE,
  shared_hp = TRUE,
  add_hp = FALSE,
  add_clust = FALSE,
  int_mu_v = c(0, log(3)),
  int_mu_l = c(log(1/300), log(1/200)),
  int_i_v = c(0, log(1.5)),
  int_i_l = c(log(1/300), log(1/200)),
  int_i_sigma = c(0, 0.2),
  m0_slope = c(-5, 5),
  m0_intercept = c(-50, 50),
  warning_grid = TRUE
)

simu_db(
  n_tasks = 10,
  n_points = 10,
  n_clusters = 1,
  n_outputs = 1,
  input2 = FALSE,
  prior_means = NULL,
  grid_input = seq(-1, 1, 0.01),
  grid_input2 = seq(-1, 1, 0.5),
  shared_input = FALSE,
  shared_hp = TRUE,
  add_hp = FALSE,
  add_clust = FALSE,
  int_mu_v = c(0, log(3)),
  int_mu_l = c(log(1/300), log(1/200)),
  int_i_v = c(0, log(1.5)),
  int_i_l = c(log(1/300), log(1/200)),
  int_i_sigma = c(0, 0.2),
  m0_slope = c(-5, 5),
  m0_intercept = c(-50, 50),
  warning_grid = TRUE
)
}
\arguments{
\item{n_tasks}{An integer. The number of tasks (per cluster).}

\item{n_points}{An integer. The number of observations per task, for each
output.}

\item{n_clusters}{An integer. The number of underlying clusters.}

\item{n_outputs}{An integer. The number of outputs.
If n_outputs = 1, a standard SE kernel is used.
If n_outputs > 1, the Multi-Output Convolution kernel structure is simulated.}

\item{input2}{A logical value indicating whether the dataset should
include an additional input named 'Input2'.}

\item{prior_means}{A vector of numbers. Prior means for each output's mean
process. If \code{NULL} (default), a random linear trend is generated for
each output (see \code{m0_intercept}/\code{m0_slope}).}

\item{grid_input}{A vector of numbers defining a grid of observations
(i.e. the reference inputs).}

\item{grid_input2}{A vector of numbers defining a grid of observations
(i.e. the reference inputs).}

\item{shared_input}{A logical value indicating whether the reference inputs
are shared across tasks.}

\item{shared_hp}{A logical value indicating whether the hyper-parameters are
shared across tasks. If TRUE and n_points >1, the hyper-parameters remain
different between the clusters.}

\item{add_hp}{A logical value indicating whether the values of
hyper-parameters should be added as columns in the dataset.}

\item{add_clust}{A logical value indicating whether the name of the
clusters should be added as a column in the dataset.}

\item{int_mu_v}{A vector of 2 numbers. Interval for mean process variance.}

\item{int_mu_l}{A vector of 2 numbers. Interval for mean process lengthscale.}

\item{int_i_v}{A vector of 2 numbers. Interval for task process variance.}

\item{int_i_l}{A vector of 2 numbers. Interval for task process lengthscale.}

\item{int_i_sigma}{A vector of 2 numbers. Interval for noise hyper-parameter.}

\item{m0_slope}{A vector of 2 numbers. Slope interval for the mean
process's default linear prior trend.}

\item{m0_intercept}{A vector of 2 numbers. Intercept interval for the mean
process's default linear prior trend.}

\item{warning_grid}{A boolean, indicating whether data generation should be
stopped by anticipation if the size of input grid is > 2000 points. To
force data generation to complete, switch this argument to \code{FALSE}}
}
\value{
A full dataset of simulated training data.
}
\description{
Simulate a complete synthetic dataset, which may be representative of various
applications. Several flexible arguments allow adjustment of the number of
tasks, clusters, outputs, or dimension of observed inputs, and the values of
many parameters controlling the data generation.
}
\examples{


data = simu_data()

data_MO = simu_data(n_outputs = 2)

data_clust = simu_data(n_clusters = 3)

data_2D = simu_data(
   input2 = TRUE,
   grid_input = seq(-1, 1, 0.1),
   grid_input2 = seq(-1, 1, 0.1))
}
