% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simu_db}
\alias{simu_db}
\alias{simu_data}
\title{Simulate a dataset tailored for MagmaClustR}
\usage{
simu_db(
  M = 10,
  N = 10,
  K = 1,
  O = 1,
  covariate = FALSE,
  prior_means = NULL,
  grid = seq(-1, 1, 0.01),
  grid_cov = seq(-1, 1, 0.01),
  common_input = TRUE,
  common_hp = TRUE,
  add_hp = FALSE,
  add_clust = FALSE,
  int_mu_v = c(0, log(3)),
  int_mu_l = c(log(1/300), log(1/200)),
  int_i_v = c(0, log(1.5)),
  int_i_l = c(log(1/300), log(1/200)),
  int_i_sigma = c(0, 0.2),
  lambda_int = c(30, 40),
  m_int = c(0, 10),
  lengthscale_int = c(30, 40),
  m0_slope = c(-5, 5),
  m0_intercept = c(-50, 50)
)

simu_data(
  M = 10,
  N = 10,
  K = 1,
  O = 1,
  covariate = FALSE,
  prior_means = NULL,
  grid = seq(-1, 1, 0.01),
  grid_cov = seq(-1, 1, 0.01),
  common_input = TRUE,
  common_hp = TRUE,
  add_hp = FALSE,
  add_clust = FALSE,
  int_mu_v = c(0, log(3)),
  int_mu_l = c(log(1/300), log(1/200)),
  int_i_v = c(0, log(1.5)),
  int_i_l = c(log(1/300), log(1/200)),
  int_i_sigma = c(0, 0.2),
  lambda_int = c(30, 40),
  m_int = c(0, 10),
  lengthscale_int = c(30, 40),
  m0_slope = c(-5, 5),
  m0_intercept = c(-50, 50)
)
}
\arguments{
\item{M}{An integer. The number of individual per cluster.}

\item{N}{An integer. The number of observations per individual.}

\item{K}{An integer. The number of underlying clusters.}

\item{O}{An integer. The number of outputs. If O = 1, standard SE kernels are used.
If O > 1, the Multi-Output Convolution kernel structure is simulated.}

\item{covariate}{A logical value indicating whether the dataset should
include an additional input covariate named 'Covariate'.}

\item{prior_means}{A vector of numbers. Prior means for the MO processes (O > 1).}

\item{grid}{A vector of numbers defining a grid of observations
(i.e. the reference inputs).}

\item{grid_cov}{A vector of numbers defining a grid of observations
(i.e. the covariate reference inputs).}

\item{common_input}{A logical value indicating whether the reference inputs
are common to all individual.}

\item{common_hp}{A logical value indicating whether the hyper-parameters are
common to all individual. If TRUE and K>1, the hyper-parameters remain
different between the clusters.}

\item{add_hp}{A logical value indicating whether the values of
hyper-parameters should be added as columns in the dataset.}

\item{add_clust}{A logical value indicating whether the name of the
clusters should be added as a column in the dataset.}

\item{int_mu_v}{A vector of 2 numbers. Interval for mean process variance (O=1).}

\item{int_mu_l}{A vector of 2 numbers. Interval for mean process lengthscale (O=1).}

\item{int_i_v}{A vector of 2 numbers. Interval for individual process variance (O=1).}

\item{int_i_l}{A vector of 2 numbers. Interval for individual process lengthscale (O=1).}

\item{int_i_sigma}{A vector of 2 numbers. Interval for noise hyper-parameter (O=1).}

\item{lambda_int}{A vector of 2 numbers. Lambda parameter of 2D exponential.}

\item{m_int}{A vector of 2 numbers. Mean of 2D exponential.}

\item{lengthscale_int}{A vector of 2 numbers. Lengthscale of 2D exponential.}

\item{m0_slope}{A vector of 2 numbers. Slope interval for m0.}

\item{m0_intercept}{A vector of 2 numbers. Intercept interval for m0.}
}
\value{
A full dataset of simulated training data.
}
\description{
Simulate a complete training dataset, which may be representative of various
applications. Several flexible arguments allow adjustment of the number of
individuals, of observed inputs, and the values of many parameters
controlling the data generation.
}
