% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simu_indiv_se}
\alias{simu_indiv_se}
\title{Simulate a batch a data}
\usage{
simu_indiv_se(ID, input, covariate, mean, v, l, sigma)
}
\arguments{
\item{ID}{An identification code, whether numeric or character.}

\item{input}{A vector of numbers. The input variable that is used as
'reference' for input and outputs.}

\item{covariate}{A vector of numbers. An additional input variable, observed
along with each reference input.}

\item{mean}{A vector of numbers. Prior mean values of the GP.}

\item{v}{A number. The variance hyper-parameter of the SE kernel.}

\item{l}{A number. The lengthscale hyper-parameter of the SE kernel.}

\item{sigma}{A number. The noise hyper-parameter.}
}
\value{
A tibble containing a batch of output data along with input and
additional information for a simulated individual.
}
\description{
Simulate a batch a output data, corresponding to one individual, coming from
a GP with a the Squared Exponential kernel as covariance structure, and
specified hyper-parameters and input.
}
\examples{
MagmaClustR:::simu_indiv_se("A", 1:10, 0, rep(0, 10), 2, 1, 0.5)
MagmaClustR:::simu_indiv_se("B", 1:10, 2:11, 3:12, 1, 1, 1)
MagmaClustR:::simu_indiv_se("C", 1:10, 5, rep(0, 10), 2, 1, 0.5)
}
