% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{sort_grid_object}
\alias{sort_grid_object}
\title{Sort a grid object}
\usage{
sort_grid_object(grid_obj)
}
\arguments{
\item{grid_obj}{A numeric vector or data.frame representing the input grid.}
}
\value{
The sorted grid object, maintaining its original structure (vector or data.frame).
}
\description{
Sorts the elements of a grid. If the grid is a data.frame, it orders the rows
based on the values of its columns (from the first to the last column).
}
\keyword{internal}
