% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swimmers}
\alias{swimmers}
\title{French swimmers performances data on 100m freestyle events}
\format{
\subsection{\code{swimmers}}{

A data frame with 76,832 rows and 4 columns:
\describe{
\item{ID}{Indentifying number associated to each swimmer}
\item{Input}{Age in years}
\item{Output}{Performance in seconds on a 100m freestyle event}
\item{Gender}{Competition gender}
}
}
}
\source{
\url{https://ffn.extranat.fr/webffn/competitions.php?idact=nat}
}
\usage{
swimmers
}
\description{
A subset of data from reported performances of French swimmers during
100m freestyle competitions between 2002 and 2016. See
https://link.springer.com/article/10.1007/s10994-022-06172-1 and
https://www.mdpi.com/2076-3417/8/10/1766 for dedicated description and
analysis.
}
\keyword{datasets}
