% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-training.R
\name{train_magma_VEM}
\alias{train_magma_VEM}
\title{Training Magma with a Variation EM algorithm}
\usage{
train_magma_VEM(
  data,
  prior_mean_k,
  ini_hp_k,
  ini_hp_i,
  kern_0,
  kern_i,
  ini_tau_i_k = NULL,
  common_hp_k = F,
  common_hp_i = F,
  pen_diag
)
}
\arguments{
\item{data}{A tibble or data frame. Columns required: \code{ID}, \code{Input}
, \code{Output}.
Additional columns for covariates can be specified.
The \code{ID} column contains the unique names/codes used to identify each
individual/task (or batch of data).
The \code{Input} column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
\code{Output} column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at each
reference \code{Input}.}

\item{prior_mean_k}{prior mean parameter of the K mean GPs (mu_k)}

\item{ini_hp_k}{named vector, tibble or data frame of hyper-parameters
associated with \code{kern_0}, the mean process' kernel. The
columns/elements should be named according to the hyper-parameters
that are used in \code{kern_0}.}

\item{ini_hp_i}{A tibble or data frame of hyper-parameters
associated with \code{kern_i}, the individual processes' kernel.
Required column : \code{ID}. The \code{ID} column contains the unique
names/codes used to identify each individual/task. The other columns
should be named according to the hyper-parameters that are used in
\code{kern_i}.}

\item{kern_0}{A kernel function, associated with the mean GP.}

\item{kern_i}{A kernel function, associated with the individual GPs.}

\item{ini_tau_i_k}{initial values of probabiliy to belong to each cluster for each individuals.}

\item{common_hp_k}{boolean indicating whether hp are common among mean GPs (for each mu_k)}

\item{common_hp_i}{boolean indicating whether hp are common among individual GPs (for each y_i)}

\item{pen_diag}{A number. A jitter term, added on the diagonal to prevent
numerical issues when inverting nearly singular matrices.}
}
\value{
A list, containing the results of the EM algorithm used for training
in Magma. The elements of the list are:
\itemize{
\item hp_k: A tibble containing the trained hyper-parameters for the mean
process' kernel.
\item hp_i: A tibble containing all the trained hyper-parameters for the
individual processes' kernels.
\item pi_k :
\item param :
\item Converged: A logical value indicated whether the EM algorithm converged
or not.
\item Training_time: Total running time of the complete training.
}
}
\description{
Training Magma with a Variation EM algorithm
}
\examples{
k = seq_len(3)
m_k <- c("K1" = 0, "K2" = 0, "K3" = 0)

db <- simu_db(N = 10, common_input = TRUE)
hp_k <- MagmaClustR:::hp("SE", list_ID = names(m_k))
hp_i <- MagmaClustR:::hp("SE", list_ID = unique(db$ID))
old_tau_i_k = MagmaClustR:::ini_tau_i_k(db = db, k = length(k), nstart = 50)

train_magma_VEM(db, m_k, hp_k, hp_i, "SE", "SE", old_tau_i_k, FALSE, FALSE, 0.1)
}
