% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_new_gp}
\alias{train_new_gp}
\title{Training new gaussian process}
\usage{
train_new_gp(db, mean_mu, cov_mu, ini_hp_i, kern_i)
}
\arguments{
\item{db}{Database with all individuals in training set}

\item{mean_mu}{mean value of mean GP at timestamps (obs + pred)}

\item{cov_mu}{covariance value of mean GP at timestamps (obs + pred)}

\item{ini_hp_i}{Initial values of each parameters of the HP to start the training.}

\item{kern_i}{Kernel associated to individual GPs.}
}
\value{
list of trained HP
}
\description{
Training new gaussian process
}
\examples{
TRUE
}
