% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-training.R
\name{train_new_gp_EM}
\alias{train_new_gp_EM}
\title{Training new GP with the EM algorithm}
\usage{
train_new_gp_EM(data, param_mu_k, ini_hp_i, kern_i, hp_i = NULL)
}
\arguments{
\item{data}{A tibble or data frame. Columns required: \code{ID}, \code{Input}
, \code{Output}.
Additional columns for covariates can be specified.
The \code{ID} column contains the unique names/codes used to identify each
individual/task (or batch of data).
The \code{Input} column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
\code{Output} column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at each
reference \code{Input}.}

\item{param_mu_k}{list of parameters for the K mean Gaussian processes}

\item{ini_hp_i}{your initial hyperparameters for your individuals}

\item{kern_i}{A kernel function, associated with the individual GPs.}

\item{hp_i}{A named vector, tibble or data frame of hyper-parameters
associated with \code{kern_i}.}
}
\value{
list of trained HP, tau_k
}
\description{
Training new GP with the EM algorithm
}
