% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{training}
\alias{training}
\title{Training}
\usage{
training(db, prior_mean, ini_hp_0, ini_hp_i, kern_0, kern_i, common_hp = T)
}
\arguments{
\item{db}{Database with all individuals in training set. Column required : 'ID', input', 'Output'}

\item{prior_mean}{prior mean parameter of the mean GP (mu_0)}

\item{ini_hp_0}{your initial hyperparameters for your kernel 0}

\item{ini_hp_i}{your initial hyperparameters for your individuals}

\item{kern_0}{kernel associated to covariance functions of the mean GP}

\item{kern_i}{kernel associated to common covariance functions of all individuals GPs}

\item{common_hp}{your common Hyperparameters}
}
\value{
list of trained HP, boolean to indicate convergence
}
\description{
Training
}
\examples{
TRUE
}
