% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-training.R
\name{update_hp_k_mixture_star_EM}
\alias{update_hp_k_mixture_star_EM}
\title{Update hp_mixture k star EM}
\usage{
update_hp_k_mixture_star_EM(db, mean_k, cov_k, kern, hp, prop_mixture_k)
}
\arguments{
\item{db}{data}

\item{mean_k}{mean}

\item{cov_k}{cov}

\item{kern}{kernel}

\item{hp}{hp}

\item{prop_mixture_k}{prop_mixture}
}
\value{
update hp_mixture star
}
\description{
Update hp_mixture k star EM
}
\examples{
TRUE
}
