% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-algorithm.R
\name{update_inv}
\alias{update_inv}
\title{Update Inv}
\usage{
update_inv(prior_inv, list_inv_i)
}
\arguments{
\item{prior_inv}{inverse of the covariance matrix of the prior mean GP (mu_0). dim = all inputs}

\item{list_inv_i}{list of inverse of the covariance matrices of each individuals. dim = inputs of i}
}
\value{
inverse of the covariance of the posterior mean GP (mu_0 | (y_i)_i). dim = (all inputs)^2
}
\description{
Update Inv
}
