% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-algorithm.R
\name{update_mean}
\alias{update_mean}
\title{update Mean}
\usage{
update_mean(prior_mean, prior_inv, list_inv_i, list_value_i)
}
\arguments{
\item{prior_mean}{mean parameter of the prior mean GP (mu_0)}

\item{prior_inv}{inverse of the covariance matrix of the prior mean GP (mu_0). dim = (all inputs)^2}

\item{list_inv_i}{list of inverse of the covariance matrices of each individuals. dim = (inputs of i)^2}

\item{list_value_i}{list of outputs (y_i) for each individuals. dim = (inputs of i) x 1list of outputs (y_i) for each individuals. dim = (inputs of i) x 1}
}
\value{
mean parameter of the posterior mean GP (mu_0 | (y_i)_i). dim = (all inputs) x 1
}
\description{
update Mean
}
