% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-training.R
\name{update_tau_star_k_EM}
\alias{update_tau_star_k_EM}
\title{Update tau star k EM}
\usage{
update_tau_star_k_EM(db, mean_k, cov_k, kern, hp, pi_k)
}
\arguments{
\item{db}{data}

\item{mean_k}{mean}

\item{cov_k}{cov}

\item{kern}{kernel}

\item{hp}{hp}

\item{pi_k}{pi}
}
\value{
update tau star
}
\description{
Update tau star k EM
}
\examples{
TRUE
}
