% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-plot-functions.R
\name{V_plot_db}
\alias{V_plot_db}
\title{Plot smoothed raw data}
\usage{
V_plot_db(db, cluster = F, legend = F)
}
\arguments{
\item{db}{database with format : ID, Input, Output}

\item{cluster}{boolean indicating whether you want cluster or not}

\item{legend}{boolean indicating whether want legend or not}
}
\value{
Graph of smoothed curves of raw data.
}
\description{
Plot smoothed raw data
}
\examples{
TRUE
}
