% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight}
\alias{weight}
\title{Weight follow-up data of children in Singapore}
\format{
\subsection{\code{weight}}{

A data frame with 3,629 rows and 4 columns:
\describe{
\item{ID}{Indentifying number associated to each child}
\item{sex}{Biological gender}
\item{Input}{Age in months}
\item{Output}{Weight in kilograms}
}
}
}
\source{
\url{https://gustodatavault.sg/}
}
\usage{
weight
}
\description{
A subset of data from the GUSTO project (https://www.gusto.sg/) collecting
the weight over time of several children in Singapore.
See https://arxiv.org/abs/2011.07866 for dedicated description and
analysis.
}
\keyword{datasets}
