% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{dmnorm}
\alias{dmnorm}
\title{Compute the Multivariate Gaussian likelihood}
\usage{
dmnorm(x, mu, inv_Sigma, log = FALSE)
}
\arguments{
\item{x}{A vector, containing values the likelihood is evaluated on.}

\item{mu}{A vector or matrix, specifying the mean parameter.}

\item{inv_Sigma}{A matrix, specifying the inverse of covariance parameter.}

\item{log}{A logical value, indicating whether we return the log-likelihood.}
}
\value{
A number, corresponding to the Multivariate Gaussian log-likelihood.
}
\description{
Modification of the function \code{dmvnorm()} from the package
\code{mvtnorm}, providing an implementation of the Multivariate Gaussian
likelihood. This version uses inverse of the covariance function as argument
instead of the traditional covariance.
}
\examples{
TRUE
}
