% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-algorithm.R
\name{e_step}
\alias{e_step}
\title{E Step}
\usage{
e_step(db, m_0, kern_0, kern_i, hp_0, hp_i, pen_diag)
}
\arguments{
\item{db}{full database with all individuals. Columns required : ID, input, Output}

\item{m_0}{mean of your Gaussian Process initial (mu_0)}

\item{kern_0}{kernel used to compute the covariance matrix of the mean GP at corresponding inputs (K_0)}

\item{kern_i}{kernel used to compute the covariance matrix of individuals GP at corresponding inputs (Psi_i)}

\item{hp_0}{hyperparameters for the kernel 0}

\item{hp_i}{hyperparameters for each kernel i}

\item{pen_diag}{value of the penalization of the diagonal}
}
\value{
mean and covariance parameters of the mean GP (mu_0)
}
\description{
E Step
}
