% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magmaClust.R
\name{full_algo_clust}
\alias{full_algo_clust}
\title{Full algorithm MAGMACLUST}
\usage{
full_algo_clust(
  data,
  new_data,
  timestamps,
  kern_i,
  ini_hp_mixture = NULL,
  common_hp_k = T,
  common_hp_i = T,
  prior_mean = NULL,
  kern_0 = NULL,
  list_hp = NULL,
  mu_k = NULL,
  ini_hp_k = NULL,
  ini_hp_i = NULL,
  hp_new_i = NULL
)
}
\arguments{
\item{data}{A tibble or data frame. Required columns: \code{ID}, \code{Input}
, \code{Output}. Additional columns for covariates can be specified.
The \code{ID} column contains the unique names/codes used to identify each
individual/task (or batch of data).
The \code{Input} column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
\code{Output} column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at
each reference \code{Input}.}

\item{new_data}{Database containing data for a new individual we want a prediction on.}

\item{timestamps}{Timestamps we want to predict at.}

\item{kern_i}{A kernel function, associated with the individual GPs. ("SE",
"PERIO" and "RQ" are aso available here)}

\item{ini_hp_mixture}{Initial probability to belong to each cluster for individuals. Optional, if list_hp is given}

\item{common_hp_k}{A boolean indicating whether hp are common among mean GPs (for each mu_k)}

\item{common_hp_i}{A boolean indicating whether hp are common among individual GPs (for each y_i)}

\item{prior_mean}{Prior arbitrary value for the mean process. Optional, not needed if 'mu' is given.}

\item{kern_0}{A kernel function, associated with the mean GP.
Several popular kernels
(see \href{https://www.cs.toronto.edu/~duvenaud/cookbook/}{The Kernel
Cookbook}) are already implemented and can be selected within the
following list:
\itemize{
\item "SE": (default value) the Squared Exponential Kernel (also called
Radial Basis Function or Gaussian kernel),
\item "LIN": the Linear kernel,
\item "PERIO": the Periodic kernel,
\item "RQ": the Rational Quadratic kernel.
Compound kernels can be created as sums or products of the above kernels.
For combining kernels, simply provide a formula as a character string
where elements are separated by whitespaces (e.g. "SE + PERIO"). As the
elements are treated sequentially from the left to the right, the product
operator '*' shall always be used before the '+' operators (e.g.
'SE * LIN + RQ' is valid whereas 'RQ + SE * LIN' is  not).
}}

\item{list_hp}{Hyper-parameters for all individuals in training set. Optional, computed if NULL.}

\item{mu_k}{list containing parameters of the mean GPs at all prediction timestamps. Optional, computed if NULL.}

\item{ini_hp_k}{named vector, tibble or data frame of hyper-parameters
associated with \code{kern_0}, the mean process' kernel. The
columns/elements should be named according to the hyper-parameters
that are used in \code{kern_0}.}

\item{ini_hp_i}{A tibble or data frame of hyper-parameters
associated with \code{kern_i}, the individual processes' kernel.
Required column : \code{ID}. The \code{ID} column contains the unique
names/codes used to identify each individual/task. The other columns
should be named according to the hyper-parameters that are used in
\code{kern_i}.#' @param hp_new_i Hyper-pameters for the new individual to predict. Optional, computed if NULL.}

\item{hp_new_i}{Hyper-pameters for the new individual to predict. Optional, computed if NULL.}
}
\value{
predicted GP parameters | posterior mean process | all trained hyperparameters


}
\description{
Full algorithm MAGMACLUST
}
