% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-elbos.R
\name{gr_clust_multi_GP}
\alias{gr_clust_multi_GP}
\title{Gradient of the elbo for a mixture of GPs}
\usage{
gr_clust_multi_GP(hp, db, hyperpost, kern, pen_diag)
}
\arguments{
\item{hp}{A tibble, data frame or named vector containing hyper-parameters.}

\item{db}{A tibble containing the values we want to compute the elbo on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{hyperpost}{List of parameters for the K mean Gaussian processes.}

\item{kern}{A kernel function.}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
The gradient of the penalised Gaussian elbo for a mixture of GPs
}
\description{
Gradient of the elbo for a mixture of GPs
}
\examples{
TRUE
}
\keyword{internal}
