% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-likelihoods.R
\name{gr_GP_mod}
\alias{gr_GP_mod}
\title{Gradient Gaussian Process modif}
\usage{
gr_GP_mod(hp, db, mean, kern, new_cov, pen_diag = NULL)
}
\arguments{
\item{hp}{set of hyper-parameter}

\item{db}{full database with all individuals. Columns required : ID, input, Output}

\item{mean}{mean of your Gaussian Process}

\item{kern}{kernel used to compute the covariance matrix}

\item{new_cov}{posterior covariance matrix of the mean GP (mu_0).}

\item{pen_diag}{value of the penalization of the diagonal}
}
\value{
Gradient of the Gaussian Process modified
}
\description{
Gradient Gaussian Process modif
}
\examples{
TRUE
}
