% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-likelihoods-MagmaClust.R
\name{gr_GP_mod_common_hp_k}
\alias{gr_GP_mod_common_hp_k}
\title{Modified common Gaussian Process for each cluster}
\usage{
gr_GP_mod_common_hp_k(hp, db, mean, kern, post_cov, pen_diag = NULL)
}
\arguments{
\item{hp}{A tibble, data frame or name vector of hyper-parameters.}

\item{db}{A tibble containing values we want to compute logL on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{mean}{A vector, specifying the mean of the GP at the reference inputs.}

\item{kern}{A kernel function used to compute the covariance matrix at corresponding timestamps.}

\item{post_cov}{posterior covariance matrix of the mean GP (mu_0). Used to compute correction term (cor_term)#'}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
gradient of modified gaussian processes for clustering
}
\description{
Modified common Gaussian Process for each cluster
}
