% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradients-elbos.R
\name{gr_GP_mod_common_hp_k}
\alias{gr_GP_mod_common_hp_k}
\title{Gradient of the penalised elbo for multiple mean GPs with common HPs}
\usage{
gr_GP_mod_common_hp_k(hp, db, mean, kern, post_cov, pen_diag)
}
\arguments{
\item{hp}{A tibble, data frame or named vector containing hyper-parameters.}

\item{db}{A tibble containing the values we want to compute the elbo on.
Required columns: Input, Output. Additional covariate columns are allowed.}

\item{mean}{A list of the k means of the GPs at union of observed timestamps.}

\item{kern}{A kernel function}

\item{post_cov}{A list of the k posterior covariance of the mean GP (mu_k).
Used to compute correction term (cor_term)}

\item{pen_diag}{A jitter term that is added to the covariance matrix to avoid
numerical issues when inverting, in cases of nearly singular matrices.}
}
\value{
The gradient of the penalised Gaussian elbo for
the sum of the k mean GPs with common HPs.
}
\description{
Gradient of the penalised elbo for multiple mean GPs with common HPs
}
\examples{
TRUE
}
\keyword{internal}
