% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{hp}
\alias{hp}
\title{Generate Initial Hyperparameters for GP Kernels}
\usage{
hp(
  kern = "SE",
  list_task_ID = NULL,
  list_output_ID = NULL,
  shared_hp_tasks = TRUE,
  shared_hp_outputs = FALSE,
  noise = FALSE,
  hp_config = NULL
)
}
\arguments{
\item{kern}{A function, or a character string indicating the chosen type of
kernel among:
\itemize{
\item "SE": the Squared Exponential kernel,
\item "LIN": the Linear kernel,
\item "PERIO": the Periodic kernel,
\item "RQ": the Rational Quadratic kernel.
Compound kernels can be created as sums or products of the above kernels.
For combining kernels, simply provide a formula as a character string
where elements are separated by whitespaces (e.g. "SE + PERIO"). As the
elements are treated sequentially from the left to the right, the product
operator '*' shall always be used before the '+' operators (e.g.
'SE * LIN + RQ' is valid whereas 'RQ + SE * LIN' is  not).
}

In case of a custom kernel function, the argument \code{list_hp} has to be
provided as well, for designing a tibble with the correct names of
hyper-parameters.}

\item{list_task_ID}{A vector of task IDs.}

\item{list_output_ID}{A vector of output IDs.}

\item{shared_hp_tasks}{If TRUE, HPs are shared across tasks.}

\item{shared_hp_outputs}{If TRUE, HPs are shared across outputs.}

\item{noise}{If TRUE, a 'noise' hyperparameter is added.}

\item{hp_config}{A tibble providing min/max bounds for HP generation, specific
to the convolution_kernel. If NULL, default bounds are used.}
}
\value{
A \code{tibble} containing the generated hyperparameters.
}
\description{
Generate Initial Hyperparameters for GP Kernels
}
\examples{
TRUE
}
