% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{hp}
\alias{hp}
\title{Generate random hyper-parameters}
\usage{
hp(kern = "SE", list_ID = NULL)
}
\arguments{
\item{kern}{A character string indicating the type of kernel among:
\itemize{
\item "SE": the Squared Exponential kernel,
\item "PERIO": the Periodic kernel,
\item "RQ": the Rational Quadratic kernel.
}}

\item{list_ID}{A vector, associating an \code{ID} value with each individual
for whom hyper-parameters are generated. If NULL (defaut) only one set of
hyper-parameters is return without the \code{ID} column.}
}
\value{
A tibble, gathering a set of hyper-parameters.
}
\description{
Generate a set of random hyper-parameters, specific to the chosen type of
kernel.
}
\examples{
MagmaClustR:::hp("PERIO")
}
