% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{hp}
\alias{hp}
\title{Generate random hyper-parameters}
\usage{
hp(kern = "SE")
}
\arguments{
\item{kern}{A character string indicating the type of kernel among:
\itemize{
\item "SE": the Squared Exponential kernel,
\item "PERIO": the Periodic kernel,
\item "RQ": the Rational Quadratic kernel.
}}
}
\value{
A tibble, gathering a set of hyper-parameters.
}
\description{
Generate a set of random hyper-parameters, specific to the chosen type of
kernel.
}
\examples{
hp("PERIOD")
}
