% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{hyperposterior_clust}
\alias{hyperposterior_clust}
\title{Compute the hyper-posterior distribution for each cluster in MagmaClust}
\usage{
hyperposterior_clust(
  data,
  mixture,
  hp_k,
  hp_i,
  kern_k,
  kern_i,
  prior_mean_k = NULL,
  grid_inputs = NULL,
  pen_diag = 1e-08
)
}
\arguments{
\item{data}{A tibble or data frame. Required columns: \code{ID}, \code{Input}
, \code{Output}. Additional columns for covariates can be specified.
The \code{ID} column contains the unique names/codes used to identify each
individual/task (or batch of data).
The \code{Input} column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
\code{Output} column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at
each reference \code{Input}.}

\item{mixture}{A tibble or data frame, indicating the mixture probabilities
in each cluster for each individual. Required column: \code{ID}.}

\item{hp_k}{A tibble or data frame of hyper-parameters
associated with \code{kern_k}.}

\item{hp_i}{A tibble or data frame of hyper-parameters
associated with \code{kern_i}.}

\item{kern_k}{A kernel function, associated with the mean GPs.
Several popular kernels
(see \href{https://www.cs.toronto.edu/~duvenaud/cookbook/}{The Kernel
Cookbook}) are already implemented and can be selected within the
following list:
\itemize{
\item "SE": (default value) the Squared Exponential Kernel (also called
Radial Basis Function or Gaussian kernel),
\item "LIN": the Linear kernel,
\item "PERIO": the Periodic kernel,
\item "RQ": the Rational Quadratic kernel.
Compound kernels can be created as sums or products of the above kernels.
For combining kernels, simply provide a formula as a character string
where elements are separated by whitespaces (e.g. "SE + PERIO"). As the
elements are treated sequentially from the left to the right, the product
operator '*' shall always be used before the '+' operators (e.g.
'SE * LIN + RQ' is valid whereas 'RQ + SE * LIN' is  not).
}}

\item{kern_i}{A kernel function, associated with the individual GPs. ("SE",
"LIN", PERIO" and "RQ" are also available here)}

\item{prior_mean_k}{A named list of vectors. The elements of the list should
be named according to the cluster they refer to. Each vector corresponds
to the evaluation of the prior mean on every \code{Input} for one cluster.
This argument can typically come from a previous use of
\code{\link{training_magmaclust}}, through the
\code{ini_args$prior_mean_k} outcome. Several other format are available
to specify this argument, see ?train_magmaclust for details.}

\item{grid_inputs}{A vector, indicating the grid of additional reference
inputs on which the mean process' hyper-posterior should be evaluated.}

\item{pen_diag}{A number. A jitter term, added on the diagonal to prevent
numerical issues when inverting nearly singular matrices.}
}
\value{
Parameters of the mean GP at timestamps chosen.
}
\description{
Compute the hyper-posterior distribution for each cluster in MagmaClust
}
\examples{
TRUE
}
