% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{ini_kmeans}
\alias{ini_kmeans}
\title{ini_kmeans}
\usage{
ini_kmeans(db, k, nstart = 50, summary = F)
}
\arguments{
\item{db}{tibble containing common Input and associated Output values to cluster}

\item{k}{number of clusters set in the kmeans}

\item{nstart}{number of initialisations to try in the kmeans}

\item{summary}{boolean}
}
\value{
Tibble of ID, true clusters, and found clusters with kmeans
}
\description{
ini_kmeans
}
