% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{ini_mixture}
\alias{ini_mixture}
\title{ini_hp_mixture}
\usage{
ini_mixture(data, k, nstart = 50)
}
\arguments{
\item{data}{database}

\item{k}{cluster}

\item{nstart}{nstart}
}
\value{
ini_hp_mixture
}
\description{
ini_hp_mixture
}
\examples{
TRUE
}
