% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{logL_GP_mod}
\alias{logL_GP_mod}
\title{Modified Log Likelihood function of the Gaussian Process}
\usage{
logL_GP_mod(hp, db, mean, kern, new_cov, pen_diag)
}
\arguments{
\item{hp}{vector or list of parameters of the kernel}

\item{db}{tibble containing values we want to compute logL on. Required columns : input, Output}

\item{mean}{mean of the GP at corresponding inputs}

\item{kern}{kernel used to compute the covariance matrix at corresponding inputs}

\item{new_cov}{posterior covariance matrix of the mean GP (mu_0). Used to compute correction term (cor_term)}

\item{pen_diag}{value of the penalization of the diagonal}
}
\value{
value of the modified Gaussian log-likelihood for one GP as it appears in the model
}
\description{
Modified Log Likelihood function of the Gaussian Process
}
\examples{
db = tibble(Input= 1:5, Output= 2:6)
mean = rep(0, 5)
hp = tibble(variance = 1, lengthscale = 0.5)
logL_GP_mod(hp, db, mean, 'SE', 0, 0.001)

}
