% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em-algorithm.R
\name{m_step}
\alias{m_step}
\title{M-step of the training procedure for Magma}
\usage{
m_step(db, old_hp, mean, cov, kern_0, kern_i, m_0, common_hp)
}
\arguments{
\item{db}{full database with all individuals. Columns required : ID, input, Output}

\item{old_hp}{the set of hyper-parameters from the previous step of the EM}

\item{mean}{mean parameter of the mean GP (mu_0), computed during the E step}

\item{cov}{covariance parameter of the mean GP (mu_0), computed during the E step}

\item{kern_0}{kernel used to compute the covariance matrix of individuals GP at corresponding inputs (Psi_i)}

\item{kern_i}{kernel used to compute the covariance matrix of the mean GP at corresponding inputs (K_0)}

\item{m_0}{prior value of the mean parameter of the mean GP (mu_0). Length = 1 or nrow(db)}

\item{common_hp}{your common Hyperparameters}
}
\value{
set of optimised hyper parameters for the different kernels of the model
}
\description{
M-step of the training procedure for Magma
}
