% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-em-algorithm.R
\name{m_step_VEM}
\alias{m_step_VEM}
\title{Variation of the M-Step of the EM algorithm}
\usage{
m_step_VEM(
  db,
  old_hp_k,
  old_hp_i,
  list_mu_param,
  kern_k,
  kern_i,
  m_k,
  common_hp_k,
  common_hp_i,
  pen_diag
)
}
\arguments{
\item{db}{A tibble or data frame. Columns required: ID, Input, Output.
Additional columns for covariates can be specified.}

\item{old_hp_k}{A named vector, tibble or data frame, containing the hyper-parameters
from the previous M-step (or initialisation) associated with the clusters.}

\item{old_hp_i}{A named vector, tibble or data frame, containing the hyper-parameters
from the previous M-step (or initialisation) associated with the
individual GPs.}

\item{list_mu_param}{List of parameters of the K mean GPs.}

\item{kern_k}{kernel used to compute the covariance matrix of the mean GP at corresponding timestamps (K_0)}

\item{kern_i}{kernel used to compute the covariance matrix of individuals GP at corresponding timestamps (Psi_i)}

\item{m_k}{prior value of the mean parameter of the mean GPs (mu_k). Length = 1 or nrow(unique(db$Input))}

\item{common_hp_k}{boolean indicating whether hp are common among mean GPs (for each mu_k)}

\item{common_hp_i}{boolean indicating whether hp are common among individual GPs (for each y_i)}

\item{pen_diag}{A number. A jitter term, added on the diagonal to prevent
numerical issues when inverting nearly singular matrices.}
}
\value{
A named list, containing the elements \code{hp_k}, a tibble
containing the hyper-parameters associated with each cluster,
\code{hp_i}, a tibble containing the hyper-parameters
associated with the individual GPs, and \code{prop_mixture_k},
a tibble containing the hyper-parameters associated with each individual,
indicating in which cluster it belongs.
}
\description{
Maximization step of the variation of the EM algorithm to compute hyper-parameters of all the
kernels involved in Magma.
}
