% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{perio_kernel}
\alias{perio_kernel}
\title{Periodic Kernel}
\usage{
perio_kernel(x, y, hp)
}
\arguments{
\item{x}{a vector/matrix}

\item{y}{a vector/matrix}

\item{hp}{a tibble of the hyperparameters}
}
\value{
The value of dot production <f(t1),f(t2)> computed from the kernel
}
\description{
Periodic Kernel
}
\examples{
kernel_period(c(1,0),c(0,1),tibble::tibble(sigma = 1, lengthscale = 0.5, period = 2))
}
