% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{perio_kernel}
\alias{perio_kernel}
\title{Periodic Kernel}
\usage{
perio_kernel(x, y, hp)
}
\arguments{
\item{x}{A vector of inputs.}

\item{y}{A vector of inputs.}

\item{hp}{A tibble containing the kernel's hyperparameters.
Required columns: 'variance', 'lengthscale', and 'scale'.}
}
\value{
The evaluation of the kernel.
}
\description{
Periodic Kernel
}
\examples{
kernel_period(
  c(1, 0), c(0, 1),
  tibble::tibble(variance = 1, lengthscale = 0.5, period = 2)
)
}
