% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_db}
\alias{plot_db}
\title{Plot smoothed curves of raw data}
\usage{
plot_db(db)
}
\arguments{
\item{db}{A data frame or tibble with format : ID, Input, Output}
}
\value{
Graph of smoothed curves of raw data.
}
\description{
Display raw data under the Magma format as smoothed curves.
}
\examples{
simu_db(M = 10) \%>\% plot_db()
}
