% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_db}
\alias{plot_db}
\title{Plot smoothed curves of raw data}
\usage{
plot_db(data, legend = FALSE)
}
\arguments{
\item{data}{A data frame or tibble with format : ID, Input, Output.}

\item{legend}{A boolean indicating whether the legend should be displayed.}
}
\value{
Graph of the raw data.
}
\description{
Display raw data under the Magma format as points and lines.
}
\examples{
## Generate a synthetic dataset
db = simu_db()
## Plot the raw data
plot_db(db)
}
