% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_db}
\alias{plot_db}
\title{Plot smoothed raw data}
\usage{
plot_db(db)
}
\arguments{
\item{db}{database with format : ID, Input, Output}
}
\value{
Visualize smoothed raw data.
}
\description{
Plot smoothed raw data
}
\examples{
TRUE
}
