% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_gp}
\alias{plot_gp}
\title{Plot Magma or GP predictions}
\usage{
plot_gp(
  pred_gp,
  x_input = NULL,
  data = NULL,
  data_train = NULL,
  prior_mean = NULL,
  y_grid = NULL,
  heatmap = F,
  prob_CI = 0.95
)
}
\arguments{
\item{pred_gp}{A tibble or data frame, typically coming from
\code{\link{pred_magma}} or \code{\link{pred_gp}} functions. Required
columns: 'Input', 'Mean', 'Var'. Additional covariate columns may be
present in case of multi-dimensional inputs.}

\item{x_input}{A vector of character strings, indicating which input should
be displayed. If NULL(default) the 'Input' column is used for the x-axis.
If providing a 2-dimensional vector, the corresponding columns are used
for the x-axis and y-axis.}

\item{data}{(Optional) A tibble or data frame. Required columns: 'Input',
'Output'. Additional columns for covariates can be specified.
The 'Input' column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
'Output' column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at
each reference 'Input'.}

\item{data_train}{(Optional) A tibble or data frame, containing the training
data of the Magma model. The data set should have the same format as the
\code{data} argument with an additional column 'ID' for identifying the
different individuals/tasks. If provided, those data are displayed as
backward colourful points (each colour corresponding to one
individual/task).}

\item{prior_mean}{(Optional) A tibble or a data frame, containing the 'Input'
and associated 'Output' prior mean parameter of the GP prediction.}

\item{y_grid}{A vector, indicating the grid of values on the y-axis for which
probabilities should be computed for heatmaps of 1-dimensional
predictions. If NULL (default), a vector of length 50 is defined, ranging
between the min and max 'Output' values contained in \code{pred_gp}.}

\item{heatmap}{A logical value indicating whether the GP prediction should be
represented as a heatmap of probabilities for 1-dimensional inputs. If
FALSE (default), the mean curve and associated 95\%CI are displayed.}

\item{prob_CI}{A number between 0 and 1 (default is 0.95), indicating the
level of the Credible Interval associated with the posterior mean curve.}
}
\value{
Visualisation of a Magma or GP prediction (optional: display data
points, training data points and the prior mean function). For 1-D inputs,
the prediction is represented as a mean curve and its associated 95\%
Credible Interval, or as a heatmap of probabilities if \code{heatmap} =
TRUE. For 2-D inputs, the prediction is represented as a heatmap, where
each couple of inputs on the x-axis and y-axis are associated
with a gradient of colours for the posterior mean values, whereas the
uncertainty is indicated by the transparency (the narrower is the 95\%CI,
the more opaque is the associated colour, and vice versa)
}
\description{
Display Magma or classic GP predictions. According to the dimension of the
inputs, the graph may be a mean curve + Credible Interval or a heatmap of
probabilities.
}
\examples{
\dontrun{
## 1-dimensional example
db <- simu_db(M = 1, covariate = FALSE)
pred_gp(db) \%>\%
  plot_gp(data = db)

## 2-dimensional example
db_2D <- simu_db(M = 1, covariate = TRUE)
pred_gp(db_2D) \%>\%
  plot_gp(data = db_2D)
}
}
