% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{plot_gp}
\alias{plot_gp}
\title{Plot Gaussian Process predictions}
\usage{
plot_gp(pred_gp, data = NULL, data_train = NULL, mean = NULL, mean_CI = F)
}
\arguments{
\item{pred_gp}{tibble coming out of the pred_gp() function, columns required : 'input', 'Mean', 'Var'}

\item{data}{tibble of observational data, columns required : 'input', 'Output' (Optional)}

\item{data_train}{tibble of training data, columns required : 'ID', input', 'Output' (Optional)}

\item{mean}{tibble of estimated value for the mean process, columns required : 'input', 'Mean' (Optional)}

\item{mean_CI}{boolean indicating whether the mean curve should be displayed along with CI95 band or not (Optional)}
}
\value{
plot the predicted curve of the GP with the 0.95 confidence interval (optional display raw data)
}
\description{
Plot Gaussian Process predictions
}
\examples{
TRUE
}
