% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{posterior_mu_k}
\alias{posterior_mu_k}
\title{Posterior mu of the cluster}
\usage{
posterior_mu_k(db, timestamps, m_k, kern_0, kern_i, list_hp)
}
\arguments{
\item{db}{matrix of data columns required ('Input', 'Output')}

\item{timestamps}{timestamps on which we want a prediction}

\item{m_k}{prior value of the mean parameter of the mean GPs (mu_k). Length = 1 or nrow(db)}

\item{kern_0}{kernel associated to the covariance function of the mean GP}

\item{kern_i}{Kernel associated to individual GPs.}

\item{list_hp}{list of your hyperparameters}
}
\value{
pamameters of the mean GP at timestamps
}
\description{
Posterior mu of the cluster
}
