% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_gp}
\alias{pred_gp}
\title{Prediction Gaussian Process}
\usage{
pred_gp(db, timestamps = NULL, mean_mu = 0, cov_mu = NULL, kern, hp)
}
\arguments{
\item{db}{tibble of data columns required ('input', 'Output')}

\item{timestamps}{timestamps on which we want a prediction}

\item{mean_mu}{mean value of mean GP at timestamps (obs + pred) (matrix dim: timestamps x 1, with Input rownames)}

\item{cov_mu}{covariance of mean GP at timestamps (obs + pred) (square matrix, with Input row/colnames)}

\item{kern}{kernel associated to the covariance function of the GP}

\item{hp}{list of hyperparameters for the kernel of the GP}
}
\value{
pamameters of the gaussian density predicted at timestamps
}
\description{
Prediction Gaussian Process
}
