% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{pred_gp_clust_animate}
\alias{pred_gp_clust_animate}
\title{Prediction animate of the Gaussian Process on the clustering}
\usage{
pred_gp_clust_animate(db, timestamps = NULL, list_mu, kern, hp)
}
\arguments{
\item{db}{A tibble or data frame. Required columns: 'Input',
'Output'. Additional columns for covariates can be specified.
The 'Input' column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
'Output' column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at
each reference 'Input'.}

\item{timestamps}{timestamps on which we want a prediction}

\item{list_mu}{List containing mean and cov of the K mean GPs.
\itemize{
\item mean   : value of mean GP at timestamps (obs + pred) (matrix dim: timestamps x 1, with Input rownames)
\item cov_mu : covariance of mean GP at timestamps (obs + pred) (square matrix, with Input row/colnames)
}}

\item{kern}{A kernel function, defining the covariance structure of the GP.
Several popular kernels
(see \href{https://www.cs.toronto.edu/~duvenaud/cookbook/}{The Kernel
Cookbook}) are already implemented and can be selected within the
following list:
\itemize{
\item "SE": (default value) the Squared Exponential Kernel (also called
Radial Basis Function or Gaussian kernel),
\item "LIN": the Linear kernel,
\item "PERIO": the Periodic kernel,
\item "RQ": the Rational Quadratic kernel.
Compound kernels can be created as sums or products of the above kernels.
For combining kernels, simply provide a formula as a character string
where elements are separated by whitespaces (e.g. "SE + PERIO"). As the
elements are treated sequentially from the left to the right, the product
operator '*' shall always be used before the '+' operators (e.g.
'SE * LIN + RQ' is valid whereas 'RQ + SE * LIN' is  not).
}}

\item{hp}{A named vector, tibble or data frame of hyper-parameters
associated with \code{kern}. The columns/elements should be named
according to the hyper-parameters that are used in \code{kern}.
List containing hyper-parameters and tau_k for the new individual.}
}
\value{
tibble of classic GP predictions but with an inscreasing number of data points considered as 'observed'
}
\description{
Prediction animate of the Gaussian Process on the clustering
}
