% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{pred_gp_clust_animate}
\alias{pred_gp_clust_animate}
\title{Prediction animate of the Gaussian Process on the clustering}
\usage{
pred_gp_clust_animate(db, timestamps = NULL, list_mu, kern, hp)
}
\arguments{
\item{db}{tibble of data columns required ('input', 'Output')}

\item{timestamps}{timestamps on which we want a prediction}

\item{list_mu}{List containing mean and cov of the K mean GPs.
\itemize{
\item mean   : value of mean GP at timestamps (obs + pred) (matrix dim: timestamps x 1, with Input rownames)
\item cov_mu : covariance of mean GP at timestamps (obs + pred) (square matrix, with Input row/colnames)
}}

\item{kern}{kernel associated to the covariance function of the GP}

\item{hp}{list of hyperparameters and tau_k for the new individual}
}
\value{
tibble of classic GP predictions but with an inscreasing number of data points considered as 'observed'
}
\description{
Prediction animate of the Gaussian Process on the clustering
}
