% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_magma}
\alias{pred_magma}
\title{Magma prediction}
\usage{
pred_magma(
  data,
  hp,
  kern = "SE",
  trained_model = NULL,
  grid_inputs = NULL,
  hyperpost = NULL,
  plot = TRUE,
  get_hyperpost = FALSE,
  get_full_cov = FALSE,
  pen_diag = 0.01
)
}
\arguments{
\item{hp}{list of hyperparameters for the kernel of the GP}

\item{kern}{kernel associated to the covariance function of the GP}

\item{grid_inputs}{grid_inputs on which we want a prediction}

\item{db}{tibble of data columns required ('input', 'Output')}

\item{mean_mu}{mean value of mean GP at grid_inputs (obs + pred) (matrix dim: grid_inputs x 1, with Input rownames)}

\item{cov_mu}{covariance of mean GP at grid_inputs (obs + pred) (square matrix, with Input row/colnames)}
}
\value{
pamameters of the gaussian density predicted at grid_inputs
}
\description{
Magma prediction
}
\examples{
db <- simu_db(M = 1, N = 10)
grid_inputs <- tibble::tibble(Input = 1:10, Covariate = 2:11)
hp <- tibble::tibble("variance" = 2, "lengthscale" = 1)
hyperpost <- list(
  "mean" = rep(0, 10),
  "cov" = kern_to_cov(union(db$Input, grid_inputs$Input), "SE", hp)
)

pred_magma(db, hp, grid_inputs = grid_inputs, hyperpost = hyperpost)
}
