% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{pred_magmaclust}
\alias{pred_magmaclust}
\title{Prediction Gaussian Process on the clustering}
\usage{
pred_magmaclust(
  data_obs,
  data_train = NULL,
  grid_inputs = NULL,
  list_mu = NULL,
  kern = "SE",
  hp_new_indiv = NULL,
  nb_cluster = NULL,
  ini_hp_k = NULL,
  ini_hp_i = NULL,
  trained_magmaclust = NULL,
  plot = TRUE,
  pen_diag = 0.01
)
}
\arguments{
\item{data_obs}{A tibble or data frame wich we want our prediction on.
Required columns: 'Input', 'Output'.
Additional columns for covariates can be specified.
The 'Input' column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
'Output' column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at
each reference 'Input'.}

\item{data_train}{A tibble or data frame wich we want our training on.
Columns required: \code{ID}, \code{Input}, \code{Output}.
Additional columns for covariates can be specified.
The \code{ID} column contains the unique names/codes used to identify each
individual/task (or batch of data).
The \code{Input} column should define the variable that is used as
reference for the observations (e.g. time for longitudinal data). The
\code{Output} column specifies the observed values (the response
variable). The data frame can also provide as many covariates as desired,
with no constraints on the column names. These covariates are additional
inputs (explanatory variables) of the models that are also observed at each
reference \code{Input}.}

\item{grid_inputs}{The grid of inputs (reference Input and covariates) values
on which the GP should be evaluated. Ideally, this argument should be a
tibble or a data frame, providing the same columns as \code{data}, except
'Output'. Nonetheless, in cases where \code{data} provides only one
'Input' column, the \code{grid_inputs} argument can be NULL (default) or a
vector. This vector would be used as reference input for prediction and if
NULL, a vector of length 500 is defined, ranging between the min and max
Input values of \code{data}.}

\item{list_mu}{List containing mean and cov of the K mean GPs.
\itemize{
\item mean   : value of mean GP at timestamps (obs + pred)
(matrix dim: timestamps x 1, with Input rownames)
\item cov_mu : covariance of mean GP at timestamps (obs + pred)
(square matrix, with Input row/colnames)
}}

\item{kern}{A kernel function, defining the covariance structure of the GP.
Several popular kernels
(see \href{https://www.cs.toronto.edu/~duvenaud/cookbook/}{The Kernel
Cookbook}) are already implemented and can be selected within the
following list:
\itemize{
\item "SE": (default value) the Squared Exponential Kernel (also called
Radial Basis Function or Gaussian kernel),
\item "LIN": the Linear kernel,
\item "PERIO": the Periodic kernel,
\item "RQ": the Rational Quadratic kernel.
Compound kernels can be created as sums or products of the above kernels.
For combining kernels, simply provide a formula as a character string
where elements are separated by whitespaces (e.g. "SE + PERIO"). As the
elements are treated sequentially from the left to the right, the product
operator '*' shall always be used before the '+' operators (e.g.
'SE * LIN + RQ' is valid whereas 'RQ + SE * LIN' is  not).
}}

\item{hp_new_indiv}{A tibble of the Trainied results of a variation
of the EM algorithm used for training in MagmaClust.
Containing columns 'theta_new' and 'hp_k_mixture'.
'theta_new' being a new set of hyperparameters with 1 individual.
'hp_k_mixture' the probability of being in the clusters.
Can be compute with the function \code{train_new_gp_EM}.
train_new_gp_EM(simu_db(M=1))}

\item{nb_cluster}{The number of clusters wanted.}

\item{ini_hp_k}{named vector, tibble or data frame of hyper-parameters
associated with \code{kern}, the mean process' kernel. The
columns/elements should be named according to the hyper-parameters
that are used in \code{kern_k}.}

\item{ini_hp_i}{A tibble or data frame of hyper-parameters
associated with \code{kern}, the individual processes' kernel.
Required column : \code{ID}. The \code{ID} column contains the unique
names/codes used to identify each individual/task. The other columns
should be named according to the hyper-parameters that are used in
\code{kern_i}.}

\item{trained_magmaclust}{A list, gathering results coming from the use of
the \code{\link{train_magmaclust}} function.}

\item{plot}{A logical value, indicating whether a plot of the results is
automatically displayed.}

\item{pen_diag}{A number. A jitter term, added on the diagonal to prevent
numerical issues when inverting nearly singular matrices.}
}
\value{
pamameters of the gaussian density predicted at timestamps
}
\description{
Prediction Gaussian Process on the clustering
}
\examples{
\donttest{
db <- simu_db()
training_test = train_magmaclust(db)
pred_magmaclust(simu_db(M=1), trained_magmaclust = training_test)
}
}
