% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{pred_max_cluster}
\alias{pred_max_cluster}
\title{Indicates the most probable cluster}
\usage{
pred_max_cluster(mixture)
}
\arguments{
\item{mixture}{A tibble or data frame containing the mixture's probabilities.}
}
\value{
A character, corresponding to the name of the most probable cluster.
}
\description{
Indicates the most probable cluster
}
\examples{
TRUE
}
