% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{pred_max_cluster}
\alias{pred_max_cluster}
\title{Prediction of the maximum of cluster}
\usage{
pred_max_cluster(hp_mixture)
}
\arguments{
\item{hp_mixture}{hp_mixture}
}
\value{
Prediction of the maximum of cluster
}
\description{
Prediction of the maximum of cluster
}
\examples{
TRUE
}
