% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V-prediction.R
\name{pred_max_cluster}
\alias{pred_max_cluster}
\title{Indicates the most probable cluster}
\usage{
pred_max_cluster(mixture)
}
\arguments{
\item{mixture}{A tibble or data frame containing mixture probabilities.}
}
\value{
A tibble, retaining only the most probable cluster. The column
\code{Cluster} indicates the the cluster's name whereas \code{Proba}
refers to its associated probability. If \code{ID} is initially
a column of \code{mixture} (optional), the function returns the most
probable cluster for all the different \code{ID} values.
}
\description{
Indicates the most probable cluster
}
\examples{
TRUE
}
