% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{regularize_data}
\alias{regularize_data}
\title{Regularize a database}
\usage{
regularize_data(
  data,
  size_grid = 30,
  scale = FALSE,
  grid_inputs = NULL,
  summarise_fct = base::mean
)
}
\arguments{
\item{data}{A tibble or data frame. Required columns: \code{ID},
\code{Output}. The \code{ID} column contains the unique names/codes used
to identify each individual/task (or batch of data). The \code{Output}
column specifies the observed values (the response variable). The data
frame can also provide as many inputs as desired, with no constraints
on the column names.}

\item{size_grid}{A number which indicates how many points each axis of the
grid must contain.}

\item{scale}{A boolean. If TRUE, the regularization is performed on
previously scaled data. If FALSE (default), the regularization is
performed on original data.}

\item{grid_inputs}{A grid of inputs on which we want to regularize data. If
NULL (default), the function create a specific grid of inputs : for each
column of input, it creates a regular sequence from the min of the input
values to the max, with a step equal to the "size_grid" parameter.}

\item{summarise_fct}{A character or a function. If same vectors of
inputs are associated with different outputs, the user can choose the new
value of output among the following functions: min, max, mean, median.
Another function can be defined ; default is "mean".}
}
\value{
A regularized database.
}
\description{
Regularize a database
}
\examples{
TRUE
}
