% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rq_kernel}
\alias{rq_kernel}
\title{Rational Quadratic Kernel}
\usage{
rq_kernel(x, y, hp, deriv = NULL)
}
\arguments{
\item{x}{A vector of inputs.}

\item{y}{A vector of inputs.}

\item{hp}{A tibble, data frame or named vector, containing the kernel's
hyperparameters. Required columns: 'variance', 'lengthscale', and 'scale'.}

\item{deriv}{A character, indicating according to which hyper-parameter the
derivative should be computed. If NULL (defaut), the function simply returns
the evaluation of the kernel.}
}
\value{
A scalar, conresponding to the evaluation of the kernel.
}
\description{
Rational Quadratic Kernel
}
\examples{
rq_kernel(
  c(1, 0), c(0, 1),
  tibble::tibble(variance = 1, lengthscale = 0.5, scale = 3)
)
}
