% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{se_kernel}
\alias{se_kernel}
\title{Squared Exponential Kernel}
\usage{
se_kernel(x, y, hp)
}
\arguments{
\item{x}{A vector of inputs.}

\item{y}{A vector of inputs.}

\item{hp}{A tibble containing the kernel's hyperparameters.
Required columns: 'variance', 'lengthscale', and 'scale'.}
}
\value{
The evaluation of the kernel.
}
\description{
Squared Exponential Kernel
}
\examples{
se_kernel(
  c(1, 0), c(0, 1),
  tibble::tibble(variance = 1, lengthscale = 0.5)
)
}
