% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simu_indiv_se}
\alias{simu_indiv_se}
\title{Simulate a batch of data}
\usage{
simu_indiv_se(ID, input, mean, v, l, sigma)
}
\arguments{
\item{ID}{An identification code, whether numeric or character.}

\item{input}{A vector of numbers. The input variable that is used as
'reference' for input and outputs.}

\item{mean}{A vector of numbers. Prior mean values of the GP.}

\item{v}{A number. The variance hyper-parameter of the SE kernel.}

\item{l}{A number. The lengthscale hyper-parameter of the SE kernel.}

\item{sigma}{A number. The noise hyper-parameter.}
}
\value{
A tibble containing a batch of output data along with input and
additional information for a simulated individual.
}
\description{
Simulate a batch of output data, corresponding to one individual, coming from
a GP with a the Squared Exponential kernel as covariance structure, and
specified hyper-parameters and input.
}
\examples{
TRUE
}
\keyword{internal}
