% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training.R
\name{train_magma}
\alias{train_magma}
\title{Training}
\usage{
train_magma(
  data,
  prior_mean = NULL,
  ini_hp_0 = NULL,
  ini_hp_i = NULL,
  kern_0 = "SE",
  kern_i = "SE",
  common_hp = T,
  grid_inputs = NULL,
  pen_diag = 0.01,
  n_loop_max = 25,
  cv_threshold = 0.001
)
}
\arguments{
\item{data}{Database with all individuals in training set. Column required :
'ID', input', 'Output'}

\item{prior_mean}{prior mean parameter of the mean GP (mu_0)}

\item{ini_hp_0}{your initial hyperparameters for your kernel 0}

\item{ini_hp_i}{your initial hyperparameters for your individuals}

\item{kern_0}{kernel associated to covariance functions of the mean GP}

\item{kern_i}{kernel associated to common covariance functions of all individuals GPs}

\item{common_hp}{your common Hyperparameters}
}
\value{
list of trained HP, boolean to indicate convergence
}
\description{
Training
}
\examples{
TRUE
}
